### 2026-02-25 #10116 @editorialbot
[PRE REVIEW]: OpenTURNS for probabilistic modeling and uncertainty quantification
**Submitting author:** @mbaudin47 (<a href="http://orcid.org/0000-0002-0450-7947">Michael Baudin</a>)
**Repository:** https://github.com/openturns/openturns
**Branch with paper.md** (empty if default branch): josspaper
**Version:** v1.26
