### 2026-02-25 #10116 @mbaudin47
> [@mbaudin47](https://github.com/mbaudin47) - your specified branch `josspaper` does not exist in the submitted repository.

Thank you for considering this proposal. This is true. The pull request is available at https://github.com/openturns/openturns/pull/2939, but the "josspaper" branch is in my fork. This is the usual way we manage the development in OpenTURNS using Github. I cannot directly create a branch on openturn's master. Would you need the merge request https://github.com/openturns/openturns/pull/2939 to be merged, so that you can see the "paper" directory within the openturns/openturns repo?
