### 2026-02-25 07b9ebe
CI: Add abi3 job
