+++ .ci_support/run_docker_coverage.sh
@@ -8,15 +8,15 @@ UID_GID=$1
 env
 
 cd /tmp
-mkdir build && cd build
-
 cmake -DCMAKE_INSTALL_PREFIX=~/.local \
       -DCMAKE_UNITY_BUILD=ON -DCMAKE_UNITY_BUILD_BATCH_SIZE=32 \
-      -DCMAKE_C_FLAGS="--coverage" -DCMAKE_CXX_FLAGS="--coverage -fuse-ld=mold" \
+      -DCMAKE_LINKER_TYPE=MOLD \
+      -DCMAKE_C_FLAGS="--coverage" -DCMAKE_CXX_FLAGS="--coverage" \
       -DSWIG_COMPILE_FLAGS="-O1" \
       -DBLA_VENDOR=Generic \
       -DUSE_HMAT=ON \
-      /io
+      -B build /io
+cd build
 make install
 OPENTURNS_NUM_THREADS=1 ctest -R pyinstallcheck --output-on-failure --timeout 200 ${MAKEFLAGS} --repeat after-timeout:2 --schedule-random
 #make tests
