### 2026-02-25 f3b7baf
CI: Add abi3 job
