+++ .github/workflows/build.yml
@@ -86,6 +86,15 @@ jobs:
       - name: Build
         run: ./utils/build_locally.sh mindeps
 
+  abi3:
+    runs-on: ubuntu-latest
+    env:
+      MAKEFLAGS: -j4
+    steps:
+      - uses: actions/checkout@v6
+      - name: Build
+        run: ./utils/build_locally.sh abi3
+
   freebsd:
     runs-on: ubuntu-latest
     env:
