### 2026-02-25 65bafbf
CI: Add abi3 job
