### 2026-02-26 4c905e2
paper/paper.md: Comply to requirements with additional sections
