### 2026-02-26 e7cc5ea
paper/paper.md: Improve Software design section
