+++ paper/paper.bib
@@ -16,8 +16,8 @@
   title = {Uncertainty on estimated magnitudes: a new approach based on a {Poisson} point process of dimension 2},
   journal = {Pure Appl. Geophys.},
   volume = {180 (10)},
-  pages = {919-933},
-  url = {https://doi.org/10.1007/s00024-022-03222-6},
+  pages = {919-933},
+  url = {https://doi.org/10.1007/s00024-022-03222-6},
 }
 
 @misc{openturns_forum,
@@ -97,4 +97,20 @@
   pages={114134},
   year={2024},
   publisher={Elsevier}
+}
+
+@manual{salome_platform,
+  title        = {{SALOME}: The Open Source Platform for Pre- and Post-Processing for Numerical Simulation},
+  author       = {{SALOME Consortium}},
+  year         = {2026},
+  url          = {https://www.salome-platform.org/},
+  note         = {Accessed: 2026-02-26}
+}
+
+@manual{pseven,
+  title        = {pSeven: Design Space Exploration and Predictive Modeling Framework},
+  author       = {{pSeven SAS}},
+  year         = {2026},
+  url          = {https://www.pseven.io/product/pseven-core/},
+  note         = {Accessed: 2026-02-26}
 }
