### 2026-02-26 e2adae9
Sample: Improve select method
