+++ python/test/t_Sample_std.py
@@ -92,6 +92,8 @@ print("sample5 (linear)=", sample5.asPoint())
 # Selection
 indices = [1, 0, 2, 1, 2]
 print("selection=", sample5.select(indices))
+indices = list(range(len(sample5)))
+print("selection(id)=", sample5.select(indices))
 
 # html conversion
 print(sample1._repr_html_())
