### 2026-02-26 53ae8c9
CMake: Compile OTPythonWrapping in C++17 mode
