### 2026-02-28 c4350bd
plot_gpr_1d.py: Simplify GPR plotting functions
Simplify the computation of the lower and upper bounds.
Use transparency.
