### 2026-02-28 d69d9e6
Refactor GPR plotting functions and improve clarity
Simplify the computation of the lower and upper bounds.
