+++ python/doc/examples/surrogate_modeling/gaussian_process_regression/plot_gpr_1d.py
@@ -48,6 +48,7 @@ Gaussian Process Regression : quick-start
 #
 import openturns as ot
 import openturns.viewer as otv
+import numpy as np
 
 
 # %%
@@ -57,7 +58,7 @@ g = ot.SymbolicFunction(["x"], ["sin(x)"])
 # %%
 # Then we define the `x_train` variable which contains the inputs of the design of experiments of the training step.
 # Then we compute the `y_train` corresponding outputs. The variable `n_train` is the size of the training design of experiments.
-x_train = ot.Sample([[x] for x in [1.0, 3.0, 4.0, 6.0, 7.9, 11.0, 11.5]])
+x_train = ot.Sample.BuildFromPoint([1.0, 3.0, 4.0, 6.0, 7.9, 11.0, 11.5])
 y_train = g(x_train)
 n_train = x_train.getSize()
 n_train
@@ -78,30 +79,60 @@ y_test = g(x_test)
 # In order to observe the function and the location of the points in the input design of experiments, we define the following function which plots the data.
 
 
-def plot_1d_data(x_data, y_data, type="Curve", legend=None, color=None, linestyle=None):
-    """Plot the data (x_data,y_data) as a Cloud/Curve"""
+def plot_1d_data(
+    x_data,
+    y_data,
+    type="Curve",
+    legend=None,
+    color=None,
+    linestyle=None,
+    pointstyle=None,
+):
+    """Plot the data (x_data, y_data) as a Cloud/Curve"""
     if type == "Curve":
-        graphF = ot.Curve(x_data, y_data)
+        graph = ot.Curve(x_data, y_data)
     else:
-        graphF = ot.Cloud(x_data, y_data)
+        graph = ot.Cloud(x_data, y_data)
     if legend is not None:
-        graphF.setLegend(legend)
+        graph.setLegend(legend)
     if color is not None:
-        graphF.setColor(color)
+        graph.setColor(color)
     if linestyle is not None:
-        graphF.setLineStyle(linestyle)
-    return graphF
+        graph.setLineStyle(linestyle)
+    if pointstyle is not None:
+        graph.setPointStyle(pointstyle)
+    return graph
+
+
+def plot_test_data(x_test, y_test):
+    """Plot the test data using a dashed gray line"""
+    return plot_1d_data(
+        x_test, y_test, legend="model", color="gray50", linestyle="dashed"
+    )
+
+
+def plot_train_data(x_train, y_train):
+    """Plot the train data using red dots"""
+    return plot_1d_data(
+        x_train,
+        y_train,
+        type="Cloud",
+        legend="train sample",
+        color="firebrick3",
+        pointstyle="o",
+    )
+
+
+def plot_gpr_data(x_test, yTestMM):
+    """Plot the GPR data using a blue curve"""
+    return plot_1d_data(x_test, yTestMM, legend="GPR", color="dodgerblue3")
 
 
 # %%
 # Here, we draw the model and the train sample.
 graph = ot.Graph("Model and Train sample", "X", "Y", True, "")
-graph.add(
-    plot_1d_data(x_test, y_test, legend="model", color="black", linestyle="dashed")
-)
-graph.add(
-    plot_1d_data(x_train, y_train, type="Cloud", legend="train sample", color="red")
-)
+graph.add(plot_test_data(x_test, y_test))
+graph.add(plot_train_data(x_train, y_train))
 graph.setLegendPosition("upper right")
 view = otv.View(graph)
 
@@ -188,19 +219,15 @@ print(gpr_result)
 
 # %%
 gprMetamodel = gpr_result.getMetaModel()
[…]
