+++ python/doc/examples/stochastic_processes/plot_field_manipulation.py
@@ -14,11 +14,11 @@ Draw a field
 # - **Case 3**: A field defined on a  1-d mesh with  2-d values,
 # - **Case 4**: A field defined on a  1-d mesh with  scalar values.
 #
-# sphinx_gallery_thumbnail_number = 2
 
 # %%
 import openturns as ot
 import openturns.viewer as otv
+# sphinx_gallery_thumbnail_number = 2
 
 # %%
 # Case 1: A field defined on a  2-d mesh with 2-d values
@@ -116,11 +116,13 @@ view = otv.View(grid)
 # -------------------------------------------------------
 # We consider a :class:`~openturns.RandomWalk` process which distribution is the normal
 # distribution of dimension 2 with zero mean and identity covariance matrix.
-# The process is defined on :math:`\cD = [0,1]^2` which is discretized by a regular grid, using
-# the class :class:`~openturns.IntervalMesher`.
+# The process is defined on :math:`\cD = [0,1]` which is discretized by a regular grid, using
+# the class :class:`~openturns.RegularGrid`.
 dist_rw = ot.Normal(2)
 origin = dist_rw.getMean()
-mesh = ot.RegularGrid(0, 1, 10000)
+N = 10000
+step = 1.0 / N
+mesh = ot.RegularGrid(0, step, N)
 rw_process = ot.RandomWalk(origin, dist_rw, mesh)
 field = rw_process.getRealization()
 
@@ -174,7 +176,10 @@ view = otv.View(grid)
 # - when the Hurst exponent is equal to 1/2, the trajectory is Brownian motion (Wiener process),
 # - when the Hurst exponent is close to 0, the process tends to a White noise process.
 #
-cov_model = ot.FractionalBrownianMotionModel(0.01, 1.5, 0.75)
+scale = 0.01
+amplitude = 1.5
+hurst_exponent = 0.75
+cov_model = ot.FractionalBrownianMotionModel(scale, amplitude, hurst_exponent)
 mesh = ot.IntervalMesher([1000]).build(ot.Interval(0.0, 1.0))
 normal_proc = ot.GaussianProcess(cov_model, mesh)
 field = normal_proc.getRealization()
@@ -189,6 +194,21 @@ grid = ot.GridLayout(1, 2)
 g_2.setTitle('A field on 1-d mesh and scalar values - Stationary process')
 grid.setGraph(0, 0, g_2)
 grid.setGraph(0, 1, g_3)
+view = otv.View(grid)
+
+# %%
+# We draw here a trajectory from a FractionalBrownianMotionModel with different Hurst exponent.
+grid = ot.GridLayout(1, 3)
+hurst_exponent_list = [0.99, 0.50, 0.001]
+for index, hurst_exponent in enumerate(hurst_exponent_list):
+    cov_model = ot.FractionalBrownianMotionModel(scale, amplitude, hurst_exponent)
+    normal_proc = ot.GaussianProcess(cov_model, mesh)
+    field = normal_proc.getRealization()
+    g = field.drawMarginal(0, True)
+    g.setTitle(r'Fractional Brownian Motion Model, Hurst exponent = ' + str(hurst_exponent))
+    g.setXTitle(r'$x_0$')
+    g.setYTitle(r'$x_1$')
+    grid.setGraph(0, index, g)
 view = otv.View(grid)
 
 # %%

[…]
