### 2026-03-03 f930041
TruncatedDistribution: Fix simplification
Closes #3114
