### 2026-03-03 a1ff814
TruncatedDistribution: Drop useless ctor
