### 2026-03-04 5049be5
Doc: Various improvements
- Improved VertexFieldToPointFunction documentation
- Added an example to DistributionTransformation
- Added a reference paper for the Kendall plot test of a copula
- Simplified the Covariance Model theory file and introduced the modelisation of the nugget factor
- Added the example plot_create_normal_process.py that shows the impact of the nugget effect on a covariance function
- Improved the already existing example plot_create_normal_process.py
- Added a paragraph on the noise modeling in the Gaussian Process Fitter
- Added the doc of the ResourceMap key of a Field
[…]
