+++ python/doc/examples/stochastic_processes/plot_field_manipulation.py
@@ -4,83 +4,212 @@ Draw a field
 """
 
 # %%
-# The objective here is to create and manipulate a field.
-# A field is the agregation of a mesh :math:`\mathcal{M}` of a domain :math:`\mathcal{D} \in \mathbb{R}^n`
-# and a sample of values in :math:`\mathbb{R}^d` associated to each vertex of the mesh.
+# In this example, we show the different ways of drawing a :class:`~openturns.Field`
+# that is a realization of a process.
 #
-# We note :math:`(\vect{t}_0, \dots, \vect{t}_{N-1})` the vertices of :math:`\mathcal{M}`
-# and :math:`(\vect{x}_0, \dots, \vect{x}_{N-1})` the associated values in :math:`\mathbb{R}^d`.
+# We consider the following fields:
 #
-# A field is stored in the :class:`~openturns.Field` object that stores the mesh and the values at each vertex of the mesh.
-# It can be built from a mesh and values or as a realization of a stochastic process.
-# sphinx_gallery_thumbnail_number = 6
+# - **Case 1**: A field defined on a  2-d mesh with 2-d values,
+# - **Case 2**: A field defined on a  2-d mesh with scalar values,
+# - **Case 3**: A field defined on a  1-d mesh with  2-d values,
+# - **Case 4**: A field defined on a  1-d mesh with  scalar values.
+#
 
 # %%
 import openturns as ot
 import openturns.viewer as otv
+# sphinx_gallery_thumbnail_number = 2
 
 # %%
-# First, define a regular 2-d mesh
-discretization = [10, 5]
-mesher = ot.IntervalMesher(discretization)
-lowerBound = [0.0, 0.0]
-upperBound = [2.0, 1.0]
-interval = ot.Interval(lowerBound, upperBound)
-mesh = mesher.build(interval)
-graph = mesh.draw()
-graph.setTitle("Regular 2-d mesh")
-view = otv.View(graph)
+# Case 1: A field defined on a  2-d mesh with 2-d values
+# ------------------------------------------------------
+# We consider a normal process defined on :math:`\cD = [0,1]^2`, with 2-d output values, defined by a
+# covariance function which is a :class:`~openturns.TensorizedCovarianceModel` built as the tensorization of
+# two  :class:`~openturns.MaternModel` covariance functions.
+# The domain :math:`\cD` is discretized by a regular grid, using the class
+# :class:`~openturns.IntervalMesher`.
+cov_model = ot.TensorizedCovarianceModel([ot.MaternModel([0.1] * 2, [0.01], 2.5)] * 2)
+mesh = ot.IntervalMesher([30] * 2).build(ot.Interval([0.0] * 2, [1.0] * 2))
+normal_proc = ot.GaussianProcess(cov_model, mesh)
+field = normal_proc.getRealization()
 
 # %%
-# Create a field as a realization of a process
-amplitude = [1.0]
-scale = [0.2] * 2
-myCovModel = ot.ExponentialModel(scale, amplitude)
-myProcess = ot.GaussianProcess(myCovModel, mesh)
-field = myProcess.getRealization()
+# We first draw the field with arrows fixed on each vertex of the mesh.
+g_1 = field.draw()
+g_1.setTitle(r'A field on 2-d mesh and 2-d values')
+g_1.setXTitle(r'$x_0$')
+g_1.setYTitle(r'$x_1$')
 
 # %%
-# Create a field from a mesh and some values
-values = ot.Normal([0.0] * 2, [1.0] * 2, ot.CorrelationMatrix(2)).getSample(
-    len(mesh.getVertices())
-)
[…]
