+++ python/doc/examples/surrogate_modeling/polynomial_chaos/plot_chaos_lognormal.py
@@ -5,21 +5,20 @@ Pitfalls in  polynomial chaos expansion due to the input distribution
 
 # %%
 #
-# In this example, we construct a polynomial
+# In this example, we build a polynomial
 # chaos expansion (refer to :ref:`functional_chaos` for more details)
 # using a countable family of polynomials orthonormal with respect to an input distribution :math:`\mu`
-# such that
+# such that :math:`\overline{\cP} \nsubseteq L^2(\mu)`. This means that
 # the functional space generated by this orthonormal polynomial family
 # is **not dense** in the Hilbert space :math:`L^2(\mu)`
-# of square-integrable functions with respect to :math:`\mu`. In other words, this
-# family is not complete.
+# of square-integrable functions with respect to :math:`\mu`. In other words,
+# it means that the polynomials orthogonal with respect to :math:`\mu` do not form a
+# complete orthogonal set.
 #
 # Therefore, there exist functions in :math:`L^2(\mu)` that are not equal to
 # their orthogonal projection onto the closure of the polynomial space.
-# Moreover, any approximation obtained by truncating this projection
-# may lead to a low-quality surrogate model (see [feller1970]_).
 #
-# This situation occurs when :math:`\mu`
+# This situation occurs for example when :math:`\mu`
 # is not characterized by the infinite sequence of its moments.
 # The :class:`~openturns.LogNormal` distribution is one such example.
 #
@@ -31,9 +30,8 @@ Pitfalls in  polynomial chaos expansion due to the input distribution
 # .. math::
 #    \model(x) = \dfrac{1}{x}.
 #
-# Let the random variable :math:`X` follow a :class:`~openturns.LogNormal` :math:`\mathcal{LN}(0,1)`
+# Let the random variable :math:`X` follow a :class:`~openturns.LogNormal` :math:`\mathcal{LN}(0,1, 0)`
 # distribution whose PDF is denoted by :math:`\mu`.
-# This distribution is not characterized by the infinite sequence of its moments.
 #
 # This particular case is discussed in [ernst2012]_.
 #
@@ -81,17 +79,9 @@ for i in range(degreeMax):
 
 # %%
 # We build a set of polynomials with increasing degrees.
-# For each polynomial, we display its value at :math:`x = 0`.
 #
-# We anticipate that the vicinity of :math:`x = 0` will cause difficulties
-# for the surrogate model of :math:`\model`, since
-# :math:`\lim_{x \rightarrow 0} \model(x) = +\infty`.
-# Given that the polynomials take *reasonable values* in the vicinity of :math:`x = 0`,
-# we expect the polynomial surrogate model to require
-# very large linear coefficients in order to reproduce high values near :math:`x = 0`,
-# which may lead to numerical cancellation issues.
-#
-# We explore polynomial degrees up to 20.
+# We explore polynomial degrees up to 20. In the following figure, only polynomials of degrees 5, 10 and
+# 20 are plotted.
 deg_list = [5, 10, 20]
 g = ot.Graph(
     "Polynomials orthonormal with respect to the LogNormal distribution", "x", "y", True
@@ -108,7 +98,15 @@ g.setLegendPosition("topleft")
 view = otv.View(g)
 
 # %%
-# Moreover, we print the value of the first polynomials at :math:`x = 0`: they have alternates signs.
+# Moreover, we print the value of the first polynomials at :math:`x = 0`: they have alternating signs.
+#
+# We anticipate that the vicinity of :math:`x = 0` will cause difficulties
+# for the surrogate model of :math:`\model`, since
+# :math:`\lim_{x \rightarrow 0} \model(x) = +\infty`.
+# Given that the polynomials take *reasonable values* in the vicinity of :math:`x = 0`,
+# we expect the polynomial surrogate model to require
[…]

[…]
