### 2026-03-05 584e9f9
ProbabilitySimRes: Deprecate getConfidenceLength
Closes #3095
