### 2026-03-05 deae1ab
ProbabilitySimRes: Deprecate getConfidenceLength
Closes #3095
