+++ python/test/t_Waarts_oblate.py
@@ -240,7 +240,7 @@ ResultMC = myMC.getResult()
 PFMC = ResultMC.getProbabilityEstimate()
 CVMC = ResultMC.getCoefficientOfVariation()
 Variance_PF_MC = ResultMC.getVarianceEstimate()
-length90MC = ResultMC.getConfidenceLength(0.90)
+length90MC = ResultMC.getProbabilityDistribution().computeBilateralConfidenceInterval(0.9)
 ott.assert_almost_equal(PFMC, 0.0869565)
 
 #
@@ -249,7 +249,7 @@ ResultLHS = myLHS.getResult()
 PFLHS = ResultLHS.getProbabilityEstimate()
 CVLHS = ResultLHS.getCoefficientOfVariation()
 Variance_PF_LHS = ResultLHS.getVarianceEstimate()
-length90LHS = ResultLHS.getConfidenceLength(0.90)
+length90LHS = ResultLHS.getProbabilityDistribution().computeBilateralConfidenceInterval(0.9)
 ott.assert_almost_equal(PFLHS, 0.126801)
 
 #
@@ -258,21 +258,21 @@ ResultDS1 = myDS1.getResult()
 PFDS1 = ResultDS1.getProbabilityEstimate()
 CVDS1 = ResultDS1.getCoefficientOfVariation()
 Variance_PF_DS1 = ResultDS1.getVarianceEstimate()
-length90DS1 = ResultDS1.getConfidenceLength(0.90)
+length90DS1 = ResultDS1.getProbabilityDistribution().computeBilateralConfidenceInterval(0.9)
 ott.assert_almost_equal(PFDS1, 0.119112)
 
 ResultDS2 = myDS2.getResult()
 PFDS2 = ResultDS2.getProbabilityEstimate()
 CVDS2 = ResultDS2.getCoefficientOfVariation()
 Variance_PF_DS2 = ResultDS2.getVarianceEstimate()
-length90DS2 = ResultDS2.getConfidenceLength(0.90)
+length90DS2 = ResultDS2.getProbabilityDistribution().computeBilateralConfidenceInterval(0.9)
 ott.assert_almost_equal(PFDS2, 0.205649)
 
 ResultDS3 = myDS3.getResult()
 PFDS3 = ResultDS3.getProbabilityEstimate()
 CVDS3 = ResultDS3.getCoefficientOfVariation()
 Variance_PF_DS3 = ResultDS3.getVarianceEstimate()
-length90DS3 = ResultDS3.getConfidenceLength(0.90)
+length90DS3 = ResultDS3.getProbabilityDistribution().computeBilateralConfidenceInterval(0.9)
 ott.assert_almost_equal(PFDS3, 0.093882)
 
 #
@@ -281,12 +281,12 @@ ResultISS = myISS.getResult()
 PFISS = ResultISS.getProbabilityEstimate()
 CVISS = ResultISS.getCoefficientOfVariation()
 Variance_PF_ISS = ResultISS.getVarianceEstimate()
-length90ISS = ResultISS.getConfidenceLength(0.90)
+length90ISS = ResultISS.getProbabilityDistribution().computeBilateralConfidenceInterval(0.9)
 ott.assert_almost_equal(PFISS, 0.130078)
 
 ResultIS = myIS.getResult()
 PFIS = ResultIS.getProbabilityEstimate()
 CVIS = ResultIS.getCoefficientOfVariation()
 Variance_PF_IS = ResultIS.getVarianceEstimate()
-length90IS = ResultIS.getConfidenceLength(0.90)
+length90IS = ResultIS.getProbabilityDistribution().computeBilateralConfidenceInterval(0.9)
 ott.assert_almost_equal(PFIS, 0.00631, 0.0, 1e-5)
