### 2026-03-05 c3bc20f
ProbabilitySimRes: Deprecate getConfidenceLength
Closes #3095
