+++ python/test/t_Waarts_25_quadratic_terms.py
@@ -167,7 +167,7 @@ ResultMC = myMC.getResult()
 PFMC = ResultMC.getProbabilityEstimate()
 CVMC = ResultMC.getCoefficientOfVariation()
 Variance_PF_MC = ResultMC.getVarianceEstimate()
-length90MC = ResultMC.getConfidenceLength(0.90)
+length90MC = ResultMC.getProbabilityDistribution().computeBilateralConfidenceInterval(0.9)
 
 #
 # LHS
@@ -175,7 +175,7 @@ ResultLHS = myLHS.getResult()
 PFLHS = ResultLHS.getProbabilityEstimate()
 CVLHS = ResultLHS.getCoefficientOfVariation()
 Variance_PF_LHS = ResultLHS.getVarianceEstimate()
-length90LHS = ResultLHS.getConfidenceLength(0.90)
+length90LHS = ResultLHS.getProbabilityDistribution().computeBilateralConfidenceInterval(0.9)
 
 #
 
@@ -384,13 +384,7 @@ print(
 print("Pf estimation = %.5e" % PFMC)
 print("Pf Variance estimation = %.5e" % Variance_PF_MC)
 print("CoV = %.5f" % CVMC)
-print("90% Confidence Interval =", "%.5e" % length90MC)
-print(
-    "CI at 90% =[",
-    "%.5e" % (PFMC - 0.5 * length90MC),
-    "; %.5e" % (PFMC + 0.5 * length90MC),
-    "]",
-)
+print("90% Confidence Interval =", length90MC)
 print(
     "************************************************************************************************"
 )
@@ -407,13 +401,7 @@ print(
 print("Pf estimation = %.5e" % PFLHS)
 print("Pf Variance estimation = %.5e" % Variance_PF_LHS)
 print("CoV = %.5f" % CVLHS)
-print("90% Confidence Interval =", "%.5e" % length90LHS)
-print(
-    "CI at 90% =[",
-    "%.5e" % (PFLHS - 0.5 * length90LHS),
-    "; %.5e" % (PFLHS + 0.5 * length90LHS),
-    "]",
-)
+print("90% Confidence Interval =", length90LHS)
 print(
     "************************************************************************************************"
 )
