### 2026-03-05 8fc2619
ProbabilitySimRes: Deprecate getConfidenceLength
Closes #3095
