+++ python/doc/examples/reliability/reliability_analysis/plot_nais.py
@@ -127,23 +127,9 @@ print("Proba NAIS = ", proba)
 print("Current coefficient of variation = ", result.getCoefficientOfVariation())
 
 # %%
-# The length of the confidence interval of level :math:`95\%` is:
-
-# %%
-length95 = result.getConfidenceLength()
-print("Confidence length (0.95) = ", result.getConfidenceLength())
-
-# %%
-# which enables to build the confidence interval:
-
-# %%
-print(
-    "Confidence interval (0.95) = [",
-    proba - length95 / 2,
-    ", ",
-    proba + length95 / 2,
-    "]",
-)
+# The confidence interval of level :math:`95\%` is:
+ci95 = result.getProbabilityDistribution().computeBilateralConfidenceInterval(0.95)
+print("Confidence interval (0.95) = ", ci95)
 
 # %%
 # You can also get the successive thresholds used by the algorithm:
