### 2026-03-05 8f5c6a9
ProbabilitySimRes: Deprecate getConfidenceLength
Closes #3095
