+++ python/doc/examples/reliability/reliability_analysis/plot_estimate_probability_importance_sampling.py
@@ -125,18 +125,9 @@ probability = result.getProbabilityEstimate()
 print("Probability = ", probability)
 
 # %%
-# In order to compute the confidence interval, we use the `getConfidenceLength` method, which returns the length of the interval.
-# In order to compute the bounds of the interval, we divide this length by 2.
-
-# %%
-alpha = 0.05
-
-# %%
-pflen = result.getConfidenceLength(1 - alpha)
-print(
-    "%.2f%% confidence interval = [%.10f,%.10f]"
-    % ((1 - alpha) * 100, probability - pflen / 2, probability + pflen / 2)
-)
+# The confidence interval of level :math:`95\%` is:
+ci95 = result.getProbabilityDistribution().computeBilateralConfidenceInterval(0.95)
+print("Confidence interval (0.95) = ", ci95)
 
 # %%
 # We can observe the convergence history of the estimate with the `drawProbabilityConvergence`
