### 2026-03-06 c307c19
Improved OrthogonalUniVariatePolynomial and OrthogonalBasis
Now the OrthogonalUniVariatePolynomial class embed the affine transformation between the given measure and its standard representative.
Introduce `UniVariateDistributionPolynomialFactory` as a replacement for the
deprecated `StandardDistributionPolynomialFactory`. The new class detects
specific orthogonal polynomial families (Hermite, Legendre, Jacobi, Laguerre,
Chebychev, etc.) for any parameterization of the corresponding distribution,
not just the standard one (e.g., Normal(0,1), Uniform(-1,1), Beta on [-1,1]).

Key changes:
- New 
[…]
