+++ python/doc/examples/surrogate_modeling/polynomial_chaos/plot_functional_chaos_graphs.py
@@ -50,7 +50,7 @@ graph = drawFamily(jacobiFamily)
 view = otv.View(graph)
 
 # %%
-laguerreFamily = ot.LaguerreFactory(2.75, 1)
+laguerreFamily = ot.LaguerreFactory(2.75)
 graph = drawFamily(laguerreFamily)
 view = otv.View(graph)
 
