+++ python/doc/user_manual/orthogonal_basis.rst
@@ -101,6 +101,7 @@ Orthogonal multivariate functions
     :template: class.rst_t
 
     experimental.FiniteOrthogonalFunctionFactory
+    experimental.UniVariateDistributionPolynomialFactory
 
 Making orthogonal multivariate functions from orthogonal univariate functions
 -----------------------------------------------------------------------------
