+++ python/src/CMakeLists.txt
@@ -558,6 +558,7 @@ ot_add_python_module(experimental experimental_module.i
                      HiGHS.i HiGHS_doc.i
                      LinearProblem.i LinearProblem_doc.i
                      FiniteOrthogonalFunctionFactory.i FiniteOrthogonalFunctionFactory_doc.i
+                     UniVariateDistributionPolynomialFactory.i UniVariateDistributionPolynomialFactory_doc.i
                     )
 ot_add_python_module(model_process model_process_module.i
                       WhiteNoise.i WhiteNoise_doc.i
@@ -810,8 +811,8 @@ ot_add_python_module(orthogonalbasis orthogonalbasis_module.i
                       OrthogonalUniVariatePolynomialFamily.i OrthogonalUniVariatePolynomialFamily_doc.i
                       OrthonormalizationAlgorithmImplementation.i OrthonormalizationAlgorithmImplementation_doc.i
                       OrthonormalizationAlgorithm.i OrthonormalizationAlgorithm_doc.i
-                      AdaptiveStieltjesAlgorithm.i AdaptiveStieltjesAlgorithm_doc.i
-                      StandardDistributionPolynomialFactory.i StandardDistributionPolynomialFactory_doc.i
+                       AdaptiveStieltjesAlgorithm.i AdaptiveStieltjesAlgorithm_doc.i
+                       StandardDistributionPolynomialFactory.i StandardDistributionPolynomialFactory_doc.i
                       OrthogonalFunctionFactory.i OrthogonalFunctionFactory_doc.i
                       OrthogonalProductPolynomialFactory.i OrthogonalProductPolynomialFactory_doc.i
                       OrthogonalBasis.i OrthogonalBasis_doc.i
