+++ lib/src/Uncertainty/Algorithm/OrthogonalBasis/CMakeLists.txt
@@ -20,6 +20,7 @@ ot_add_source_file (OrthogonalUniVariatePolynomialFamily.cxx)
 ot_add_source_file (OrthonormalizationAlgorithm.cxx)
 ot_add_source_file (OrthonormalizationAlgorithmImplementation.cxx)
 ot_add_source_file (AdaptiveStieltjesAlgorithm.cxx)
+ot_add_source_file (UniVariateDistributionPolynomialFactory.cxx)
 ot_add_source_file (StandardDistributionPolynomialFactory.cxx)
 ot_add_source_file (OrthogonalBasis.cxx)
 ot_add_source_file (OrthogonalUniVariateFunctionFamily.cxx)
@@ -35,6 +36,7 @@ ot_install_header_file (OTOrthogonalBasis.hxx)
 ot_install_header_file (OrthogonalFunctionFactory.hxx)
 ot_install_header_file (OrthogonalUniVariatePolynomialFactory.hxx)
 ot_install_header_file (OrthogonalProductPolynomialFactory.hxx)
+ot_install_header_file (UniVariateDistributionPolynomialFactory.hxx)
 ot_install_header_file (StandardDistributionPolynomialFactory.hxx)
 ot_install_header_file (OrthogonalBasis.hxx)
 ot_install_header_file (OrthogonalUniVariatePolynomial.hxx)
