### 2026-03-06 de86423
Improved and fixed [Linear|Quadratic]LeastSquares
Now the data are centered before to compute the parameters using ordinary least-squares. It fixes #3068.
A factor 0.5 was used for the off-diagonal terms of the quadratic part. It should not. Fixed.
Adapted the other classes to take into account these changes. By the way, ParetoFactory
has been improved (no use of computeEmpiricalCDF when the sample has no tie). It fixes #3120.
