+++ python/test/t_QuadraticLeastSquares_std.py
@@ -8,13 +8,10 @@ X = [[1.0], [2.0], [3.0], [4.0]]
 Y = [[5.0], [3.0], [4.0], [7.0]]
 algo = ot.QuadraticLeastSquares(X, Y)
 algo.run()
-mm = algo.getMetaModel()
+mm = algo.getResult().getMetaModel()
 x = [0.0]
 y2 = mm(x)
 print(f"mm(x)={y2}")
-print(f"constant={algo.getConstant()}")
-print(f"linear={algo.getLinear()}")
-print(f"quadratic={algo.getQuadratic()}")
 ott.assert_almost_equal(mm.gradient([0.0])[0, 0], -5.55)
 ott.assert_almost_equal(mm.hessian([0.0])[0, 0, 0], 2.5)
 
@@ -32,14 +29,13 @@ y = f(x)
 y2 = mm(x)
 print(f"f(x)={y}")
 print(f"mm(x)={y2}")
-print(f"constant={algo.getConstant()}")
-print(f"linear={algo.getLinear()}")
-print(f"quadratic={algo.getQuadratic()}")
 ott.assert_almost_equal(y, y2, 1e-2, 1e-2)
-linear = algo.getLinear()
-ott.assert_almost_equal(linear[0, 0], 4.0, 1e-2, 1e-2)
-ott.assert_almost_equal(linear[1, 0], 7.0, 1e-2, 1e-2)
-quadratic = algo.getQuadratic()
-ott.assert_almost_equal(0.5 * quadratic[0, 0, 0], 9.0, 1e-2, 1e-2)
-ott.assert_almost_equal(0.5 * quadratic[0, 1, 0], 0.1, 1e-2, 1e-2)
-ott.assert_almost_equal(0.5 * quadratic[1, 1, 0], 6.0, 1e-2, 1e-2)
+g = f.gradient(x)
+g2 = mm.gradient(x)
+ott.assert_almost_equal(g[0, 0], g2[0, 0], 1e-2, 1e-2)
+ott.assert_almost_equal(g[1, 0], g2[1, 0], 1e-2, 1e-2)
+h = f.hessian(x)
+h2 = mm.hessian(x)
+ott.assert_almost_equal(h[0, 0, 0], h2[0, 0, 0], 1e-2, 1e-2)
+ott.assert_almost_equal(h[0, 1, 0], h2[0, 1, 0], 1e-2, 1e-2)
+ott.assert_almost_equal(h[1, 1, 0], h2[1, 1, 0], 1e-2, 1e-2)
