+++ paper/paper.md
@@ -1,0 +1,250 @@
+---
+title: 'OpenTURNS: A Python package for uncertainty quantification'
+tags:
+  - Python
+  - uncertainty quantification
+  - probabilistic modeling
+  - statistics
+  - meta-modeling
+  - calibration
+  - inverse problems
+  - stochastic processes
+  - Bayesian modeling
+  - machine learning
+  - supervised machine learning
+  - unsupervised machine learning
+  - surrogate modelling
+  - polynomial chaos expansion
+  - gaussian process regression
+  - sensitivity analysis
+  - regression
+  - function approximation
+  - Quasi-Monte Carlo methods
+  - optimization
+  - Reliability Analysis
+  - Markov chain Monte Carlo
+  - copula
+  - dimensionality reduction
+  - maximum likelihood estimation
+  - importance sampling
+authors:
+  - name: Michaël Baudin
+    orcid: 0000-0002-0450-7947
+    affiliation: 1
+  - name: Anne Dutfoy
+    affiliation: 1
+  - name: Régis Lebrun
+    affiliation: 2
+  - name: Julien Schueller
+    orcid: 0009-0006-2452-7861
+    affiliation: 3
+  - name: Sofiane Haddad
+    affiliation: 2
+  - name: Loïc Brevault
+    affiliation: 4
+  - name: Mathieu Balesdent
+    affiliation: 4
+  - name: Julien Pelamatti
+    affiliation: 1
+  - name: Joseph Muré
+    affiliation: 1
+affiliations:
+ - name: EDF R&D, France
+   index: 1
+ - name: Airbus, France
+   index: 2
+ - name: Phimeca, France
+   index: 3
+ - name: ONERA, France
+   index: 4
+date: 30 July 2025
+bibliography: paper.bib
+---
+
+# Summary
+
+OpenTURNS (www.openturns.org) is an open-source Python package for 
+probabilistic modeling, uncertainty quantification (UQ), and machine learning.
+It provides a unified interface, granting users access to a comprehensive 
+set of state-of-the-art algorithms and modeling tools.
+The library is supported by extensive, regularly updated
+documentation, including introductory examples, various use cases, 
+tutorials, and user guides, enhancing its accessibility.
+OpenTURNS integrates seamlessly with NumPy, Pandas, and other
+scientific Python packages, allowing users to develop UQ
+workflows and generate reproducible results.
+Its modular architecture enables both newcomers and advanced users to
+access reliable, well-tested methods for uncertainty propagation,
+sensitivity analysis, and reliability assessment.
+Its active community provides additional support and resources through 
+a Discourse forum [@openturns_forum] and a website.
+Stable releases are published on a semiannual basis, and the library is
+distributed through Conda, pip, and Debian repositories and available on 
+Linux, Windows, and macOS.
+
+# Statement of need
+
+UQ plays a critical role in robust decision-making during the design
+and operation of complex systems.
+It encompasses probabilistic modeling (multivariate distributions,
+stochastic processes), sampling, functional algebra, uncertainty
+propagation, sensitivity analysis, reliability, optimization, inverse
+problems, and surrogate modeling.
+Validated and well-documented methods are essential for teaching, academic 
+research, and industrial applications.
+However, many probabilistic modeling libraries are
+domain-specific—focusing on fields such as mechanical engineering, computational
+fluid dynamics, or hydrology—which limits the sharing of best practices.
+This underscores the need for a general-purpose, cross-disciplinary
+framework for probabilistic modeling.
+
+# State of the field
+
+Other Python packages exist for Monte Carlo
+sampling and other numerical methods [@virtanen2020scipy],
[…]
