+++ paper/paper.bib
@@ -1,0 +1,116 @@
+@Inbook{baudin2016,
+  author="Baudin, Micha{\"e}l and Dutfoy, Anne and Iooss, Bertrand and Popelin, Anne-Laure",
+  title="{OpenTURNS}: An Industrial Software for Uncertainty Quantification in Simulation. In: Ghanem R., Higdon D., Owhadi H. (eds) Handbook of Uncertainty Quantification.",
+  year="2016",
+  publisher="Springer International Publishing",
+  address="Cham",
+  pages="1--38",
+  isbn="978-3-319-11259-6",
+  doi="10.1007/978-3-319-11259-6_64-1",
+  url="https://doi.org/10.1007/978-3-319-11259-6_64-1"
+}
+
+@article{dutfoy2023,
+  author = {A. Dutfoy},
+  year = {2023},
+  title = {Uncertainty on estimated magnitudes: a new approach based on a {Poisson} point process of dimension 2},
+  journal = {Pure Appl. Geophys.},
+  volume = {180 (10)},
+  pages = {919-933},
+  url = {https://doi.org/10.1007/s00024-022-03222-6},
+}
+
+@misc{openturns_forum,
+  title        = {{OpenTURNS}'s forum},
+  howpublished = {\url{https://openturns.discourse.group/}},
+}
+
+@misc{persalys,
+  title = {Persalys: Graphical User Interface for Uncertainty Quantification},
+  author = {{EDF} and {Phimeca Engineering}},
+  year = {2016},
+  howpublished = {\url{https://www.persalys.fr/}}
+}
+
+@article{pedregosa2011scikit,
+  title={{Scikit-learn}: {Machine} learning in {Python}},
+  author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
+  journal={the Journal of machine Learning research},
+  volume={12},
+  pages={2825--2830},
+  year={2011},
+  publisher={JMLR. org}
+}
+
+@inproceedings{seabold2010statsmodels,
+  title={statsmodels: Econometric and statistical modeling with {Python}},
+  author={Seabold, Skipper and Perktold, Josef},
+  booktitle={9th Python in Science Conference},
+  year={2010},
+}
+
+@article{Wicaksono2023,
+  doi = {10.21105/joss.05671},
+  url = {https://doi.org/10.21105/joss.05671}, 
+  year = {2023}, publisher = {The Open Journal}, 
+  volume = {8}, number = {90}, pages = {5671}, 
+  author = {Wicaksono, Damar and Hecht, Michael}, 
+  title = {UQTestFuns: A Python3 library of uncertainty quantification (UQ) test functions}, 
+  journal = {Journal of Open Source Software}
+}
+
+@article{roy2023quasi,
+  title={Quasi-monte carlo methods in python},
+  author={Roy, Pamphile and Owen, Art B. and Balandat, Maximilian and Haberland, Matt},
+  journal={Journal of Open Source Software},
+  volume={8},
+  number={84},
+  pages={5309},
+  year={2023}
+}
+
+@article{virtanen2020scipy,
+  author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and others},
+  title   = {{SciPy} 1.0: Fundamental Algorithms for Scientific Computing in {Python}},
+  journal = {Nature Methods},
+  volume  = {17},
+  pages   = {261--272},
+  year    = {2020},
+  doi     = {10.1038/s41592-019-0686-2},
+  url     = {https://doi.org/10.1038/s41592-019-0686-2}
+}
+
[…]
