### 2026-03-09 2bb913c
paper/paper.bib: Fixed DOIs
paper/paper.md: Added libraries which depend on OpenTURNS
