+++ paper/paper.bib
@@ -1,13 +1,13 @@
 @Inbook{baudin2016,
-  author="Baudin, Micha{\"e}l and Dutfoy, Anne and Iooss, Bertrand and Popelin, Anne-Laure",
-  title="{OpenTURNS}: An Industrial Software for Uncertainty Quantification in Simulation. In: Ghanem R., Higdon D., Owhadi H. (eds) Handbook of Uncertainty Quantification.",
-  year="2016",
-  publisher="Springer International Publishing",
-  address="Cham",
-  pages="1--38",
-  isbn="978-3-319-11259-6",
-  doi="10.1007/978-3-319-11259-6_64-1",
-  url="https://doi.org/10.1007/978-3-319-11259-6_64-1"
+  author={Baudin, Micha{\"e}l and Dutfoy, Anne and Iooss, Bertrand and Popelin, Anne-Laure},
+  title={{OpenTURNS: An Industrial Software for Uncertainty Quantification in Simulation. In: Ghanem R., Higdon D., Owhadi H. (eds) Handbook of Uncertainty Quantification.}},
+  year={2016},
+  publisher={Springer International Publishing},
+  address={Cham},
+  pages={1--38},
+  isbn={978-3-319-11259-6},
+  doi={10.1007/978-3-319-11259-6_64-1},
+  url={https://doi.org/10.1007/978-3-319-11259-6_64-1}
 }
 
 @article{dutfoy2023,
@@ -15,27 +15,30 @@
   year = {2023},
   title = {Uncertainty on estimated magnitudes: a new approach based on a {Poisson} point process of dimension 2},
   journal = {Pure Appl. Geophys.},
-  volume = {180 (10)},
+  volume = {180},
+  number = {10},
   pages = {919-933},
   url = {https://doi.org/10.1007/s00024-022-03222-6},
+  doi = {10.1007/s00024-022-03222-6}
 }
 
 @misc{openturns_forum,
   title        = {{OpenTURNS}'s forum},
   howpublished = {\url{https://openturns.discourse.group/}},
+  year = {2020}
 }
 
 @misc{persalys,
-  title = {Persalys: Graphical User Interface for Uncertainty Quantification},
+  title = {{Persalys: Graphical User Interface for Uncertainty Quantification}},
   author = {{EDF} and {Phimeca Engineering}},
   year = {2016},
   howpublished = {\url{https://www.persalys.fr/}}
 }
 
 @article{pedregosa2011scikit,
-  title={{Scikit-learn}: {Machine} learning in {Python}},
+  title={{Scikit-learn: Machine learning in Python}},
   author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
-  journal={the Journal of machine Learning research},
+  journal={The Journal of Machine Learning Research},
   volume={12},
   pages={2825--2830},
   year={2011},
@@ -43,10 +46,12 @@
 }
 
 @inproceedings{seabold2010statsmodels,
-  title={statsmodels: Econometric and statistical modeling with {Python}},
+  title={{Statsmodels: Econometric and statistical modeling with Python}},
   author={Seabold, Skipper and Perktold, Josef},
-  booktitle={9th Python in Science Conference},
+  booktitle={{9th Python in Science Conference}},
   year={2010},
+  doi = {10.25080/majora-92bf1922-011},
+  url = {https://doi.org/10.25080/majora-92bf1922-011}
 }
 
 @article{Wicaksono2023,
@@ -55,23 +60,25 @@
   year = {2023}, publisher = {The Open Journal}, 
   volume = {8}, number = {90}, pages = {5671}, 
[…]
