+++ paper/paper.bib
@@ -23,6 +23,7 @@
 @misc{openturns_forum,
   title        = {{OpenTURNS}'s forum},
   howpublished = {\url{https://openturns.discourse.group/}},
+  year = {2020}
 }
 
 @misc{persalys,
@@ -113,4 +114,25 @@
   year         = {2026},
   url          = {https://www.pseven.io/product/pseven-core/},
   note         = {Accessed: 2026-02-26}
+}
+
+@manual{gemseo,
+  title        = {{GEMSEO}: Generic Engine for Multidisciplinary Scenarios, Exploration and Optimization},
+  author       = {{IRT Saint Exupéry}},
+  year         = {2026},
+  url          = {https://gemseo.readthedocs.io/},
+  note         = {Accessed: 2026-02-26}
+}
+
+@article{Roy2018,
+  doi = {10.21105/joss.00493},
+  url = {https://doi.org/10.21105/joss.00493},
+  year = {2018},
+  publisher = {The Open Journal},
+  volume = {3},
+  number = {21},
+  pages = {493},
+  author = {Pamphile T. Roy and Sophie Ricci and Romain Dupuis and Robin Campet and Jean-Christophe Jouhaud and Cyril Fournier},
+  title = {{BATMAN}: Statistical analysis for expensive computer codes made easy},
+  journal = {Journal of Open Source Software}
 }
