### 2026-03-09 37d385e
paper/paper.bib: Fixed DOIs
paper/paper.md: Added libraries which depend on OpenTURNS
