### 2026-03-12 09b495e
Sphinx: Set autodoc_use_legacy_class_based
