+++ python/doc/examples/surrogate_modeling/gaussian_process_regression/plot_gpr_noise.py
@@ -128,8 +128,8 @@ dimension = 1
 basis = ot.ConstantBasisFactory(dimension).build()
 
 # %%
-# Case 1:   Ignore any noise
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~
+# Case 1: Ignore any noise
+# ~~~~~~~~~~~~~~~~~~~~~~~~
 # We introduce no noise here: it means that we assume that the output values are exact.
 # We plot the surrogate model: we see that it interpolates the data set, as expected.
 covarianceModel = ot.MaternModel([1.0] * dimension, 1.5)
@@ -442,8 +442,8 @@ graph_traj_nuggetF.add(plot_1d_data(x_train, y_train_homosc, type="Cloud", legen
 view = otv.View(graph_traj_nuggetF)
 
 # %%
-# To facilitate comparison, we draw the trajectories generated by the Gaussian process built with a nugget factor
-# or with a noise.
+# To facilitate comparison, we draw the trajectories generated by the Gaussian process built with nugget factor
+# or with noise.
 grid = ot.GridLayout(1, 2)
 grid.setGraph(0, 0, graph_traj_nuggetF)
 grid.setGraph(0, 1, graph_traj_homosc)
