+++ python/doc/examples/surrogate_modeling/gaussian_process_regression/plot_gpr_noise.py
@@ -1,6 +1,6 @@
 """
-Gaussian Process Regression: nugget effect and  noise
-=====================================================
+Gaussian Process Regression: nugget effect and noise
+====================================================
 """
 
 # %%
