### 2026-03-16 1241b70
KDTree: Fix limit case
Closes #2617
