+++ python/test/t_KDTree_std.py
@@ -46,3 +46,16 @@ if ot.PlatformInfo.HasFeature("nanoflann"):
         assert (i in indices) == (x.norm() < radius), f"{x.norm()}"
         if i in indices:
             ott.assert_almost_equal(x.norm(), distance[indices.index(i)])
+
+# limit case
+sample = [[0.0, -1.0 / 3.0],
+          [-0.5, -1.0 / 3.0],
+          [0.5, -0.77777777777],
+          [-0.75, -0.111111111],
+          [0.25, 0.55555555555],
+          [-0.25, -0.55555555555]]
+tree = ot.KDTree(sample)
+x = sample[1]
+indices = tree.queryK(x, 6, True)
+print(indices)
+assert indices == [1, 5, 3, 0, 2, 4]
