### 2026-03-17 cfbc57f
LevelSetMesher: Avoid degenerate simplices
