### 2026-03-18 d5af5c9
Improved the LeastSquares theory doc and LinearLeastSquares and QuadraticLeastSquares API doc
