+++ python/doc/theory/numerical_methods/least_squares_resolution.rst
@@ -1,4 +1,4 @@
-.. _least_squares:
+.. _least_squares_resolution:
 
 Least squares problems numerical methods
 ----------------------------------------
@@ -9,15 +9,15 @@ Least squares problems numerical methods
   interest, or when one wishes to perform a statistical regression.
 
 | Given a matrix :math:`\matr{\Psi}~\in~\Rset^{N\times P}`, :math:`N>P`,
-  and a vector :math:`\underline{y}~\in~\Rset^{N}`, we want to find a
-  vector :math:`\underline{a}~\in \Rset^{P}` such that
-  :math:`\matr{\Psi}\: \underline{a}` is the best approximation to
-  :math:`\underline{y}` in the least squares sense. Mathematically
+  and a vector :math:`\vect{y}~\in~\Rset^{N}`, we want to find a
+  vector :math:`\vect{a}~\in \Rset^{P}` such that
+  :math:`\matr{\Psi}\: \vect{a}` is the best approximation to
+  :math:`\vect{y}` in the least squares sense. Mathematically
   speaking, we want to solve the following minimization problem:
 
   .. math::
 
-    \min_{\underline{a}} \, \, = \, \, \left\| \; \matr{\Psi} \, \underline{a} \, - \, \underline{y} \; \right\|_2
+    \min_{\vect{a}} \, \, = \, \, \left\| \; \matr{\Psi} \, \vect{a} \, - \, \vect{y} \; \right\|_2
 
   In the following, it is assumed that the rank of matrix
   :math:`\matr{\Psi}` is equal to :math:`P`.
@@ -32,7 +32,7 @@ Least squares problems numerical methods
 
   .. math::
 
-      \matr{\Psi}^{\mbox{\scriptsize \textsf{T}}} \; \matr{\Psi} \; \underline{a} \, \, = \, \, \matr{\Psi}^{\mbox{\scriptsize \textsf{T}}} \; \underline{y}
+      \matr{\Psi}^{\mbox{\scriptsize \textsf{T}}} \; \matr{\Psi} \; \vect{a} \, \, = \, \, \matr{\Psi}^{\mbox{\scriptsize \textsf{T}}} \; \vect{y}
 
 | The matrix
   :math:`\matr{C} \equiv \matr{\Psi}^{\mbox{\scriptsize \textsf{T}}} \; \matr{\Psi}`
@@ -51,13 +51,13 @@ Least squares problems numerical methods
   .. math::
 
      \begin{aligned}
-         \matr{R}^{\mbox{\scriptsize \textsf{T}}} \; \underline{z} \, \, = \, \, \matr{\Psi}^{\mbox{\scriptsize \textsf{T}}} \; \underline{y}
-         \qquad , \qquad \matr{R} \; \underline{a} \, \, = \, \, \underline{z}
+         \matr{R}^{\mbox{\scriptsize \textsf{T}}} \; \vect{z} \, \, = \, \, \matr{\Psi}^{\mbox{\scriptsize \textsf{T}}} \; \vect{y}
+         \qquad , \qquad \matr{R} \; \vect{a} \, \, = \, \, \vect{z}
        \end{aligned}
 
 | It has to be noted that this theoretical approach is seldom used in
   practice though. Indeed the resulting least squares solution is quite
-  sensitive to a small change in the data (i.e. in :math:`\underline{y}`
+  sensitive to a small change in the data (i.e. in :math:`\vect{y}`
   and :math:`\matr{\Psi}`). More precisely, the normal equations are
   always more badly conditioned than the original overdetermined system,
   as their condition number is squared compared to the original problem:
@@ -88,7 +88,7 @@ Least squares problems numerical methods
 
   .. math::
 
[…]
