### 2026-03-18 14437f2
Improved the LeastSquares theory doc and LinearLeastSquares and QuadraticLeastSquares API doc
