+++ python/doc/theory/meta_modeling/least_squares.rst
@@ -1,0 +1,166 @@
+.. _least_squares:
+
+Least squares meta models
+-------------------------
+
+A least squares meta model provides an approximation of the model :math:`\model: \Rset^\inputDim \rightarrow \Rset^\outputDim` which is valid over its whole domain of definition.
+
+Let :math:`\model: \Rset^\inputDim \rightarrow \Rset^\outputDim`. The objective is to create a meta model
+:math:`\metaModel` defined by:
+
+  .. math::
+      :label: LSMetaModel
+      \metaModel(\vect{x})  = \sum_{j=0}^N \vect{a}_j \Psi_j(\vect{x})
+
+where, for :math:`0 \leq j \leq N`, :math:`\vect{a}_j \in \Rset^\outputDim` and
+:math:`\psi_j: \Rset^\inputDim \rightarrow \Rset` with :math:`\psi_0: \vect{x} \rightarrow 1`.
+
+The functional basis :math:`(\psi_j)_{0 \leq j \leq N}` is specified and the objective it to determine the
+coefficients :math:`(\vect{a}_j)_{0 \leq j \leq N}`.
+
+Let :math:`\mat{A} \in \cM_{N+1, \outputDim}` and :math:`\vect{\Psi(\vect{x})} \in \Rset^{N+1}` be defined by:
+
+  .. math::
+    :label: defAPsi
+    \mat{A} & = \Tr{(\vect{a}_0 | \dots | \vect{a}_N)}\\
+    \vect{\Psi(\vect{x})} & = \Tr{(\Psi_0(\vect{x}), \dots, \Psi_N(\vect{x})}
+
+Then the meta model :eq:`LSMetaModel` can be written as:
+
+  .. math::
+      :label: LSMetaModelMat
+      \vect{y}  =  \Tr{\mat{A}}\vect{\Psi(\vect{x})}
+
+Let :math:`\cX` be an experimental design of size :math:`\sampleSize`, that is, a set of observations of
+the input vector defined by:
+
+.. math::
+
+    \cX = \left\{ \vect{x}^{(1)}, \dots, \vect{x}^{(\sampleSize)} \right\},
+
+as well as the corresponding output vectors:
+
+.. math::
+
+    \cY = \left\{ \vect{y}^{(1)}, \dots, \vect{y}^{(\sampleSize)} \right\}.
+
+where :math:`\vect{y}^{(k)} = \model{ \vect{x}^{(k)}}`.
+
+We define the objective function :math:`\cJ` by:
+
+.. math::
+    :label: objFunc
+    \cJ(\mat{A}) = \sum_{i=1}^\sampleSize \left \| \vect{y}^{(i)} - \Tr{\mat{A}} \vect{\Psi(\vect{x}^{(i))}
+    \right \|_{L^2}^2
+
+Then we search :math:`\widehat{\mat{A}}` that minimizes the objective function :eq:`objFunc`:
+
+.. math::
+    :label: defA
+    \widehat{\mat{A}} = \argmin \cJ(\mat{A})
+
+Let the matrices :math:`\mat{Psi} \in \cM_{\sampleSize, N+1}` and :math:`\mat{Y}\in \cM_{\sampleSize, \outputDim}`
+be defined by:
+
+.. math::
+    \mat{Psi} & = (\psi_j(\vect{x}^{(i)}))_{i,j}\\
+    \mat{Y}    & = \Tr{(\vect{y}^{(1)} | \dots | \vect{y}^{(\sampleSize)}}
+
+Then :math:`\widehat{\mat{A}}` is solution of the linear system:
+
+.. math::
+:label: pbLeastSquares
+    \Tr{\mat{Psi}}\mat{Psi}\widehat{\mat{A}} = \Tr{\mat{Psi}}\mat{Y}
+
[…]
