### 2026-03-18 ba81400
Improved the LeastSquares theory doc and LinearLeastSquares and QuadraticLeastSquares API doc
