### 2026-03-18 0f4979e
Improved QuadraticLeastSquares: thoery and API
