### 2026-03-20 5ecf1b4
GPR: Improve noise examples
