+++ python/doc/examples/surrogate_modeling/gaussian_process_regression/plot_gpr_noise.py
@@ -1,44 +1,66 @@
 """
-Gaussian Process Regression: heteroskedastic noise
-==================================================
+Gaussian Process Regression: nugget effect and noise
+====================================================
 """
 
 # %%
-# The goal of this example is to estimate a Gaussian process regression accounting for a noise (known noise).
+# In this example, we show how to estimate a Gaussian Process Regression surrogate model
+# (refer to :ref:`gaussian_process_regression`)
+# from a noisy data set of the model. We consider the following cases:
 #
-# It assumes a basic knowledge of Gaussian Process Regression.
-# For that purpose, we create a Gaussian Process Regression surrogate model for a function which has scalar real inputs and outputs.
+# - Case 1: No noise is considered, assuming that the output values are exact,
+# - Case 2: We consider a homoscedastic noise on the output values,
+# - Case 3: We consider a heteroscedastic noise on the output values,
+# - Case 4: We consider a homoscedastic noise modeled as a nugget factor.
+#
+# Both Case 2 and Case 4 model a homoscedastic noise on the output values of the model. But the two models
+# are quite different:
+#
+# - the :meth:`~openturns.GaussianProcessFitter.setNoise` method of the :class:`~openturns.GaussianProcessFitter`
+#   takes noise into account only in
+#   the expression of the model likelihood, in the evaluation of the covariance parameters,
+# - the :meth:`~openturns.CovarianceModel.setNuggetFactor` method of the :class:`~openturns.CovarianceModel`
+#   modifies the covariance model of
+#   the Gaussian process and keeps it modified once the parameters have been estimated. As a result, the
+#   trajectories of the Gaussian process become less smooth than with the first model.
+#
+# Modeling a noise on the output values as a nugget effect is interesting to
+# simulate some sensors which have a finite precision and generates some output values perturbated by
+# a White Noise process.
+#
+# On the contrary, taking noise into account with the :meth:`~openturns.GaussianProcessFitter.setNoise` method
+# impacts the estimation of the covariance parameters
+# but not the regularity of the process.
 
 # %%
 # Introduction
 # ------------
 #
-# We consider the sine function:
+# We consider the model :math:`\model: [0,12] \rightarrow \Rset` defined by:
 #
 # .. math::
-#    y = x \sin(x)
+#
+#      y = \model (x) = x \sin(x)
 #
 #
-# for any :math:`x\in[0,12]`.
-#
-# We want to create a surrogate of this function using GPR:
-#
-# .. math::
-#    y_k=x_k \sin(x_k)
-#
-# We will consider a known centered Gaussian observation noise:
+# We want to create a surrogate of this model using Gaussian Process Regression, from the data set defined by:
 #
 # .. math::
-#   \widehat{\vect{y}}_k
-#    = \vect{y}_k + \epsilon_k, \quad \epsilon_k \sim \mathcal{N}\left(0, \tau_k^2\right)
+#
+#      y_k = x_k \sin(x_k), \quad 1 \leq k \leq \sampleSize
+#
 
 # %%
 import openturns as ot
+import openturns.experimental as otexp
 import openturns.viewer as otv
+import math as m
 
 # %%
 # First let us introduce some useful function.
[…]
