### 2026-03-20 2489faa
Deprecate KrigingAlgorithm related classes
This just officially deprecates them as they were already removed from the doc in v1.26

Closes #3053
