### 2026-03-24 2242b0e
Revise matrix representation in Gaussian process regression
Updated matrix notation for clarity and consistency.
