### 2026-03-24 600c7aa
gaussian_process_regression.rst: Fixed equations
Revise matrix representation in Gaussian process regression for clarity and consistency.
Simplify LaTeX representation of mean vector
Updated the covariance matrix equation to include noise terms.
