+++ python/doc/theory/meta_modeling/gaussian_process_regression.rst
@@ -49,13 +49,11 @@ where:
 
 .. math::
 
-    \vect{\mu}(\vect{x}) = \left(
-      \begin{array}{l}
-        \mu_1(\vect{x}) \\
-        \vdots  \\
-        \mu_\outputDim(\vect{x})
-       \end{array}
-     \right)
+   \vect{\mu}(\vect{x}) = \begin{pmatrix}
+      \mu_1(\vect{x}) \\
+      \vdots \\
+      \mu_\outputDim(\vect{x})
+    \end{pmatrix}
 
 with :math:`\mu_\ell(\vect{x}) = \sum_{j=1}^{b} \beta_j^\ell \varphi_j(\vect{x})` and
 :math:`\varphi_j: \Rset^\inputDim \rightarrow \Rset` the trend functions basis for :math:`1 \leq j \leq b` and
@@ -88,35 +86,28 @@ We note:
 
 .. math::
 
-    \vect{y} = \left(
-      \begin{array}{l}
+    \vect{y} = \begin{pmatrix}
         \vect{y}_1 \\
-        \vdots  \\
+        \vdots \\
         \vect{y}_{\sampleSize}
-       \end{array}
-     \right) \in \Rset^{\outputDim \times \sampleSize},
-     \quad
-     \vect{m}_{\vect{\beta}} = \left(
-      \begin{array}{l}
+    \end{pmatrix} \in \Rset^{\outputDim \times \sampleSize},
+    \quad
+    \vect{m}_{\vect{\beta}} = \begin{pmatrix}
         \vect{\mu}(\vect{x}_1) \\
-        \vdots  \\
+        \vdots \\
         \vect{\mu}(\vect{x}_{\sampleSize})
-       \end{array}
-     \right) \in \Rset^{\outputDim \times \sampleSize}
-
+    \end{pmatrix} \in \Rset^{\outputDim \times \sampleSize}
 
 and:
 
 .. math::
     :label: CovaMatDef
 
-    \mat{C}_{\vect{p}} = \left(
-      \begin{array}{lcl}
-        \mat{C}_{11} & \dots &  \mat{C}_{1 \times \sampleSize}\\
-        \vdots &   & \vdots \\
-        \mat{C}_{\sampleSize \times 1} & \dots &  \mat{C}_{\sampleSize \times \sampleSize}
-       \end{array}
-     \right) \in \cS_{\outputDim \times \sampleSize}^+(\Rset)
+    \mat{C}_{\vect{p}} = \begin{pmatrix}
+      \mat{C}_{11} & \dots & \mat{C}_{1 \times \sampleSize} \\
+      \vdots & & \vdots \\
+      \mat{C}_{\sampleSize \times 1} & \dots & \mat{C}_{\sampleSize \times \sampleSize}
+    \end{pmatrix} \in \cS_{\outputDim \times \sampleSize}^+(\Rset)
 
 where :math:`\mat{C}_{ij} = C_{\vect{p}}(\vect{x}_i, \vect{x}_j)\in \cS_{\outputDim \times \outputDim}^+
 (\Rset)`.
@@ -181,35 +172,29 @@ This step is performed by the class :class:`~openturns.GaussianProcessFitter`.
 
 **Add a noise to the output values**: We show how to take into account the fact that the output values
 of the function are not known precisely. This noise is modeled by normal distribution with zero mean and a
-covariance matrix :math:`\mat{\Sigma}_i^{noise} \in \cM_{\outputDim \times \outputDim}(\Rset)` that can be
+covariance matrix :math:`\mat{\Sigma}_i^{\text{noise}} \in \cM_{\outputDim \times \outputDim}(\Rset)` that can be
[…]
