### 2026-03-24 f82d2c7
Simplify LaTeX representation of mean vector
