### 2026-03-24 1a81bc8
Refactor mathematical notation in Gaussian process doc
