### 2026-03-24 07a649e
FCA: Add domination method flag
See #3099
