### 2026-03-24 6ebaadb
FCA: Add domination method flag
See #3099

Closes #3161
