### 2026-03-25 970624a
Apply suggestions from code review
Co-authored-by: Michael Baudin <31351465+mbaudin47@users.noreply.github.com>
