### 2026-03-26 5daa86b
CalibrationResult: Drop log-scale logic
