### 2026-03-26 43bc688
Update gaussian_process_regression.rst
