### 2026-03-26 01d524e
Fix formatting in Gaussian process regression documentation
