### 2026-03-26 65cf992
Fix typo in Gaussian process regression documentation
