### 2026-03-27 a7f1d34
Apply suggestions from code review
Co-authored-by: Michael Baudin <31351465+mbaudin47@users.noreply.github.com>
