### 2026-03-27 f738029
Fix formatting of mathematical notation in documentation
