### 2026-03-27 c23fe93
Add noise covariance to Gaussian process regression
Updated the covariance matrix equation to include noise terms.
