### 2026-04-01 e6b2f4d
GaussianProcessFitter: Fixed link to the theory help page
