### 2026-04-01 f39a0df
Promoted the use of \vect in doc
