+++ python/doc/examples/data_analysis/sample_analysis/plot_linear_regression.py
@@ -6,14 +6,14 @@ Build and validate a linear model
 # %%
 # In this example, we  build a linear regression model and validate it numerically and graphically.
 #
-# The linear model links a scalar variable :math:`Y` and to an n-dimensional one :math:`\underline{X} = (X_i)_{i \leq n}`, as follows:
+# The linear model links a scalar variable :math:`Y` and to an n-dimensional one :math:`\vect{X} = (X_i)_{i \leq n}`, as follows:
 #
 # .. math::
 #    \tilde{Y} = a_0 + \sum_{i=1}^n a_i X_i + \varepsilon
 #
 # where :math:`\varepsilon` is the residual, supposed to follow :math:`\mathcal{N}(0.0, 1.0)`.
 #
-# The linear model may be validated graphically if :math:`\underline{X}` is of dimension 1, by drawing on the same graph the cloud :math:`(X_i, Y_i)`.
+# The linear model may be validated graphically if :math:`\vect{X}` is of dimension 1, by drawing on the same graph the cloud :math:`(X_i, Y_i)`.
 #
 # The linear model can also be validated numerically with several tests:
 #
@@ -24,7 +24,7 @@ Build and validate a linear model
 #
 # The hypothesis on the residuals (centered Gaussian distribution) may be validated:
 #
-# - graphically if :math:`\underline{X}` is of dimension 1, by drawing the residual couples (:math:`\varepsilon_i, \varepsilon_{i+1}`),
+# - graphically if :math:`\vect{X}` is of dimension 1, by drawing the residual couples (:math:`\varepsilon_i, \varepsilon_{i+1}`),
 #   where the residual :math:`\varepsilon_i` is evaluated on the samples :math:`(X, Y)`.
 # - numerically with the `LinearModelResidualMean` test which tests, under the hypothesis of a Gaussian sample, if the mean of the residual is equal to zero.
 #   It is based on the Student test (equality of mean for two Gaussian samples).
