### 2026-04-01 3817436
Promoted the use of \vect in doc
