### 2026-04-01 cdb0f55
Doc: Mention Spack
