### 2026-04-01 5553d61
Improved Mesh::draw3D
Now, the colors are controlled through ResourceMap entries, allowing for the use of alpha channel. One can also set a color per simplex.
Some default values have been modified to give a better default rendering.
