### 2026-04-01 57a68e6
Improved Mesh::draw3D
Now, the colors are controlled through ResourceMap entries, allowing for the use of alpha channel.
Some default values have been modified to give a better default rendering.
