+++ lib/etc/openturns.conf.in
@@ -427,9 +427,9 @@
   <DiracCovarianceModel-CheckUnique value_bool="true"         />
 
   <!-- OT::SpectralModelImplementation parameters -->
-  <SpectralModel-DefaultMaximumFrequency value_float="5.0"           />
-  <SpectralModel-DefaultMinimumFrequency value_float="-5.0"          />
-  <SpectralModel-DefaultFrequencyNumber  value_int="129" />
+  <SpectralModel-DefaultMaximumFrequency value_float="5.0"  />
+  <SpectralModel-DefaultMinimumFrequency value_float="-5.0" />
+  <SpectralModel-DefaultFrequencyNumber  value_int="129"    />
 
   <!-- OT::FieldImplementation parameters -->
   <Field-AutomaticScaling value_bool="true"  />
@@ -438,29 +438,36 @@
   <Field-LevelNumber      value_int="10"     />
 
   <!-- OT::SampleImplementation parameters -->
-  <Sample-CSVFileSeparator      value_str=";"     />
-  <Sample-CSVPrecision           value_int="16" />
+  <Sample-CSVPrecision           value_int="16"         />
+  <Sample-PrintEllipsisSize      value_int="3"          />
+  <Sample-PrintEllipsisThreshold value_int="1000"       />
+  <Sample-SmallKendallTau        value_int="23"         />
+  <Sample-CommentMarker          value_str="#"          />
+  <Sample-CSVFileSeparator       value_str=";"          />
   <Sample-CSVFormat              value_str="scientific" />
-  <Sample-CommentMarker          value_str="#"    />
-  <Sample-PrintEllipsisSize      value_int="3"    />
-  <Sample-PrintEllipsisThreshold value_int="1000" />
-  <Sample-PrintFormat            value_str="{:.7g}" />
-  <Sample-SmallKendallTau        value_int="23"   />
+  <Sample-PrintFormat            value_str="{:.7g}"     />
 
   <!-- OT::DomainImplementation parameters -->
-  <Domain-SmallVolume value_float="1.0e-12" />
-  <Domain-DefaultInColor value_str="prettyOrange" />
-  <Domain-DefaultOutColor value_str="prettyBlue" />
+  <Domain-SmallVolume     value_float="1.0e-12"    />
+  <Domain-DefaultInColor  value_str="prettyOrange" />
+  <Domain-DefaultOutColor value_str="prettyBlue"   />
 
   <!-- OT::Mesh parameters -->
-  <Mesh-BackfaceCulling      value_bool="false"   />
-  <Mesh-AmbientFactor        value_float="0.1"    />
-  <Mesh-DiffuseFactor        value_float="0.7"    />
-  <Mesh-Shininess            value_float="100.0"  />
-  <Mesh-SpecularFactor       value_float="0.2"    />
-  <Mesh-VertexEpsilon        value_float="1.0e-6" />
-  <Mesh-LargeSize            value_int="5000"     />
-  <Mesh-CheckValidity        value_bool="false"   />
+  <Mesh-BackfaceCulling      value_bool="false"    />
+  <Mesh-CheckValidity        value_bool="false"    />
+  <Mesh-AmbientFactor        value_float="0.1"     />
+  <Mesh-DiffuseFactor        value_float="0.7"     />
+  <Mesh-Shininess            value_float="100.0"   />
+  <Mesh-SpecularFactor       value_float="0.2"     />
+  <Mesh-VertexEpsilon        value_float="1.0e-6"  />
+  <Mesh-DefaultThetaX        value_float="0.3"     />
+  <Mesh-DefaultThetaY        value_float="0.3"     />
+  <Mesh-DefaultThetaZ        value_float="0.3"     />
[…]
