### 2026-04-01 ac851e9
Improved Functional Chaos Expansion doc
Closes #3160
