+++ python/doc/examples/surrogate_modeling/polynomial_chaos/plot_enumeratefunction.py
@@ -1,13 +1,21 @@
 """
-Plot enumeration rules
-----------------------
+Plot enumeration function
+-------------------------
 """
 
 # %%
-# In order to build up a functional chaos multivariate basis :math:`\{\psi_{\idx},\idx \in \NM\}`
-# by tensorization of univariate basis terms, we need to enumerate the multi-indices :math:`\vect{\alpha} \in \mathbb{N}^{n_X}`.
-# In this example we are going to explore properties of these enumeration rules.
-# Refer also to :ref:`enumeration_strategy` in the theoric documentation.
+# This example illustrates the enumeration functions which are ijections between :math:`\Nset` into :math:`\Nset^\inputDim`.
+# Refer to :ref:`enumeration_strategy` to get the precise description of the enumerate functions. We detail here the bijections:
+#
+# - Linear enumeration function
+# - Hyperbolic enumeration function
+# - Anisotropic hyperbolic enumeration function
+# - Infinity norm enumeration function
+#
+# These bijections are used in the in the :ref:`functional chaos expansion setting <functional_chaos>`.
+# In order to build up a multivariate basis :math:`\{\psi_{\vect{\alpha}},\vect{\alpha} \in \Nset^\inputDim\}`
+# by tensorization of univariate basis terms, we need to enumerate the multi-indices :math:`\vect{\alpha} \in \Nset^\inputDim`.
+# In this example, we interprete the impact of the different enumeration functions within the functional chaos expansion setting.
 
 import openturns as ot
 import openturns.viewer as otv
@@ -15,32 +23,33 @@ import math as m
 
 # %%
 # The simplest way to generate the multi-indices is to enumerate the terms of increasing length.
-# In other words, we enumerate the multi-indices with length equal to 0, then 1, 2, 3, etc.
-# This is called "graded reverse-lexicographic ordering" in [sullivan2015]_.
-# This is named the linear enumeration rule in the library; let us instantiate it in the 2-dimensional case.
-dim = 2
+# If the basis is polynomial, then the length corresponds to the total degree of the polynomial.
+# Within a strata, the multi-indices are ordered according to the "graded reverse lexicographic
+# ordering" in [sullivan2015]_.
+# This is named the linear enumeration rule in the library.
+#
+# We print the ordered elements in dimension 4 to illustrate the ordering of the multi-indices.
+# We only consider the 4 first strata and we print their associated length
+# (which are equal to 0 up to 3).
+dim = 4
 enum_func = ot.LinearEnumerateFunction(dim)
-
-# %%
-# Print the 25 first multi-indices and their associated length.
 td_prev = 0
-print("#  | multi-index | length")
-print("---+-------------+-------------")
-for i in range(25):
+print("#  | multi-index     | length")
+print("---+-----------------+-------------")
+i_max = enum_func.getStrataCumulatedCardinal(3)
+for i in range(i_max):
     multi_index = enum_func(i)
     td = sum(multi_index)
     if td > td_prev:
         td_prev = td
-        print("---+-------------+-------------")
+        print("---+-----------------+-------------")
     print(f"{i:2} |       {multi_index} |           {td}")
 
 # %%
-# Plot the multi-indices of the enumeration rule by stratas.
-# In the specific case of the linear enumerate function, each strata contains
[…]
