+++ utils/lint.sh
@@ -7,7 +7,7 @@ set -e
 
 PATH=~/.local/bin/:$PATH flake8 python/ validation --ignore=W503,E203 --max-line-length=185
 
-PATH=~/.local/bin/:$PATH doc8 python/doc/ --ignore D000 --ignore-path python/doc/theory/meta_modeling/orthogonal_polynomials.rst --ignore-path python/doc/theory/reliability_sensitivity/sensitivity_sobol.rst --max-line-length 185
+PATH=~/.local/bin/:$PATH doc8 python/doc/ --ignore D000 --ignore-path python/doc/theory/meta_modeling/orthonormal_polynomials.rst --ignore-path python/doc/theory/reliability_sensitivity/sensitivity_sobol.rst --max-line-length 185
 
 # check for windows line endings
 ! find python lib -type f | xargs file | grep CRLF
