### 2026-04-01 cbff792
Improved Functional Chaos Expansion doc
Closes #3160
