+++ python/doc/examples/surrogate_modeling/polynomial_chaos/plot_chaos_ishigami.py
@@ -1,6 +1,6 @@
 """
-Create a polynomial chaos for the Ishigami function: a quick start guide to polynomial chaos
-============================================================================================
+Quick start:  the Ishigami function
+===================================
 """
 
 # %%
