+++ python/doc/examples/surrogate_modeling/polynomial_chaos/plot_chaos_ishigami_dependent_input.py
@@ -1,0 +1,182 @@
+"""
+Create a FCE for dependent inputs: transformation vs domination
+===============================================================
+"""
+
+# %%
+# In this example, we create a functional chaos expansion for the
+# :ref:`Ishigami function<use-case-ishigami>` when the input distribution
+# has dependent marginals.
+#
+# Refer to :ref:`functional_chaos` to learn more about functional chaos expansion.
+#
+# We provide one input sample and one output sample of the Ishigami function. We build two meta models:
+#
+# - Meta model 1: we use an isoprobabilistic transformation that maps the input distribution to another one with independent marginals.
+# - Meta model 2: we use the domination method: the basis of the projection space is not orthonormal to the input distribution.
+#
+
+
+# %%
+# Define the Ishigami model
+# -------------------------
+
+# %%
+from openturns.usecases import ishigami_function
+import openturns as ot
+import openturns.viewer as otv
+
+# %%
+# We load the Ishigami model. The `IshigamiModel` data class contains the input
+# distribution :math:`\mu_{\inputRV}` of the random vector :math:`\vect{X}=(X_1, X_2, X_3)`
+# in `im.distribution`.
+im = ishigami_function.IshigamiModel()
+input_names = im.distribution.getDescription()
+
+# %%
+# We want to introduce some dependence betwenn the components. That is why we have to change the
+# input distribution stored in the Ishigami model (which have independent components).
+# We use a copula
+# that links :math:`(X_1,X_2)` with a :class:`~openturns.ClaytonCopula`. The last component
+# :math:`X_3` is independent of :math:`(X_1,X_2)`. The final 3d-copula is a
+# :class:`~openturns.BlockIndependentCopula` copula.
+# We keep the initial marginal distributions.
+copula = ot.BlockIndependentCopula([ot.ClaytonCopula(1.5), ot.IndependentCopula(1)])
+input_dist = ot.JointDistribution([im.distribution.getMarginal(i) for i in range(im.dim)], copula)
+
+# %%
+# We generate an input sample from the input distribution and
+# we compute the output sample from the Ishigami function that is contained
+# in `im.model`.
+sampleSize = 1000
+inputTrain = im.distribution.getSample(sampleSize)
+outputTrain = im.model(inputTrain)
+
+# %%
+# We display the relationships between the outputs and the inputs.
+grid = ot.VisualTest.DrawPairsXY(inputTrain, outputTrain)
+view = otv.View(grid, figure_kw={"figsize": (12.0, 4.0)})
+
+# %%
+# We draw the histogramm of the output values.
+graph = ot.HistogramFactory().build(outputTrain).drawPDF()
+graph.setTitle("Ishigami outputs")
+graph.setXTitle("y")
+view = otv.View(graph)
+
+# %%
+# Meta model 1: Transformation method
+# -----------------------------------
+# As the input distribution has dependent marginals, we can use an
+# :ref:`isoprobabilistic transformation <isoprobabilistic_transformation>` that maps the
+# input distribution to
+# another one with independent marginals. This new distribution is not specified, therefore the
+# default adaptive strategy and the default projection stategy are used.
+#
+# The default adaptive strategy is defined as follows:
+#
+# - a basis is built as the tensorization of the univariate
+#   polynomials family
+#   orthonormal to the standard representative of the input marginals distribution. Then
+#   this basis is
+#   orthonormal to the distribution denoted by :math:`\tilde{\mu}` which is the tensorization
+#   of the standard representative of the input marginals distribution;
+# - the enumerate function is chosen according to
[…]
