+++ python/doc/examples/surrogate_modeling/polynomial_chaos/plot_functional_chaos_graphs.py
@@ -1,6 +1,6 @@
 """
-Polynomial chaos graphs
-=======================
+Visualize orthonormal polynomials basis
+=======================================
 """
 
 # %%
