### 2026-04-01 49b6ab1
Improved Functional Chaos Expansion doc
