### 2026-04-01 c46ebad
Improved Functional Chaos Expansion doc
Closes #3160
