+++ python/doc/examples/surrogate_modeling/polynomial_chaos/plot_functional_chaos_advanced_ctors.py
@@ -1,6 +1,6 @@
 """
-Advanced polynomial chaos construction
-======================================
+Use advanced features for PCE
+=============================
 """
 
 # %%
