+++ python/doc/theory/meta_modeling/orthonormal_polynomials.rst
@@ -1,70 +1,33 @@
-.. _orthogonal_polynomials:
+.. _orthonormal_polynomials:
 
-Orthogonal polynomials
-----------------------
+Univariate orthonormal polynomials
+----------------------------------
 
-This page provides mathematical details on sequences of
-orthogonal polynomials. Some of these sequences will be used to
-construct the basis of the so-called *polynomial chaos expansion*.
+This page provides mathematical details on sequences of univariate
+polynomials which are orthonornal with respect to a distribution :math:`\mu`.
 
-Mathematical framework
-~~~~~~~~~~~~~~~~~~~~~~
+These sequences are used to
+build multivariate polynomial basis by tensorization in the polynomial chaos expansion (refer to
+:ref:`functional_chaos` and :ref:`chaos_basis`).
 
-Orthogonal polynomials are associated to an inner product, defined
-as follows.
-Given an *interval of orthogonality* :math:`[\alpha,\beta]`
-(:math:`\alpha \in \Rset \cup \{-\infty\}`,
-:math:`\beta \in \Rset \cup \{\infty\}`, :math:`\alpha < \beta`) and a
-weight function :math:`w(x)> 0`, the polynomials :math:`P`
-and :math:`Q` are orthogonal if:
-
-.. math::
-
-    \scalarproduct{P}{Q} = \int_{\alpha}^{\beta}P(x)Q(x)~w(x) dx = 0
-
-Therefore, a sequence of orthogonal polynomials :math:`(P_n)_{n\geq 0}`
-(:math:`P_n`: polynomial of degree :math:`n`) verifies:
-
-.. math::
-
-    \scalarproduct{P_m}{P_n} = 0 \text{~~for every~~} m \neq n
-
-The chosen inner product induces a norm on polynomials in the usual
-way:
-
-.. math::
-
-    \parallel P_n\parallel = \scalarproduct{P_n}{P_n}^{1/2}
-
-In the following, we consider weight functions :math:`w(x)`
-corresponding to *probability density functions*, which satisfy:
-
-.. math::
-
-    \int_{\alpha}^{\beta} \; w(x) \;  dx \, \, = \,\, 1
-
-Moreover, we consider families of *orthonormal* polynomials
-:math:`(P_n)_{n\geq 0}`, that is polynomials with a unit norm:
-
-.. math::
-
-    \|P_n\| \, \, = \, \, 1
-
-Any sequence of orthogonal polynomials has a recurrence formula
-relating any three consecutive polynomials as follows:
+A sequence of univariate orthonormal polynomials is defined by a three-term recurrence
+formula. Let by :math:`P_n` the polynomial of degree :math:`n`.
+Then the sequence :math:`(P_n)_{n \in \Nset}` verifies the relation:
 
 .. math::
 
     P_{n+1}\ =\ (a_nx+b_n)\ P_n\ +\ c_n\ P_{n-1}
 
-Orthogonormal polynomials with respect to usual probability distributions
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+We detail:
 
-Below is a table showing an example of particular (normalized)
-orthogonal polynomials associated with *continuous* weight functions.
-Note that the orthonormal polynomials are
-orthonormal with respect to the standard representative distribution
-of the given distribution.
+- some well-known sequences of univariate orthonormal polynomials,
+- how to compute any sequence of univariate orthonormal polynomials.
+
+Orthonormal polynomials with respect to usual probability distributions
[…]
