+++ python/doc/theory/meta_modeling/functional_chaos.rst
@@ -3,320 +3,317 @@
 Functional Chaos Expansion
 --------------------------
 
-Introduction
-~~~~~~~~~~~~
-
 Accounting for the joint probability density function (PDF)
-:math:`\mu_{\vect{X}}(\vect{x})` of the input random vector
-:math:`\vect{X}`, one seeks the joint PDF of output random vector
-:math:`\vect{Y} = g(\vect{X})`. This may be achieved using
+:math:`\mu_{\inputRV}(\vect{x})` of the input random vector
+:math:`\inputRV`, one seeks the joint PDF of output random vector
+:math:`\outputRV = \model(\inputRV)`. This may be achieved using
 Monte Carlo (MC) simulation (see :ref:`monte_carlo_simulation`). However, the MC
 method may require a large number of model evaluations, i.e. a great
 computational cost, in order to obtain accurate results.
 
 A possible solution to overcome this problem is to project the model
-:math:`g` in a suitable functional space, such as
-the Hilbert space :math:`L^2(\mu_{\vect{X}})` of square-integrable functions with
-respect to :math:`\mu_{\vect{X}}`.
-More precisely, we may consider an expansion of the model onto an orthonormal
-basis of :math:`L^2(\mu_{\vect{X}})`.
-As an example of this type of expansion, one can mention expansions by
-wavelets, polynomials, etc.
+:math:`\model` in a suitable functional space, such as
+the Hilbert space :math:`L^2\left(\mu_{\inputRV}\right)` of square-integrable functions with
+respect to :math:`\mu_{\inputRV}`. This projection is called a *functional chaos expansion* of
+the model :math:`\model`.
 
 The principles of the building of a functional chaos expansion are described in the sequel.
 
-Model
-~~~~~
+Methodology: principles
+~~~~~~~~~~~~~~~~~~~~~~~
 
 We consider the output random vector:
 
 .. math::
 
-    \vect{Y} = g(\vect{X})
+    \outputRV = \model(\inputRV)
 
-where :math:`g: \Rset^{n_X} \rightarrow \Rset^{n_Y}` is the model,
-:math:`\vect{X}` is the input random vector which distribution is
-:math:`\mu_{\vect{X}}`,
-:math:`n_X \in \Nset` is the input dimension,
-:math:`n_Y \in \Nset` is the output dimension.
-We assume that :math:`\vect{Y}` has finite variance i.e.
-:math:`g\in L^2(\mu_{\vect{X}})`.
+where :math:`\model: \Rset^{\inputDim} \rightarrow \Rset^{\outputDim}` is the model,
+:math:`\inputRV` is the input random vector which distribution is
+:math:`\mu_{\inputRV}`.
 
-When :math:`n_Y > 1`, the functional chaos algorithm is used on each marginal
-of :math:`\vect{Y}`, using the same multivariate orthonormal basis for
+We assume that :math:`\outputRV` has finite variance i.e.
+:math:`\model \in L^2\left(\mu_{\inputRV}\right)`.
+
+When :math:`\outputDim > 1`, the functional chaos algorithm is used on each marginal
+of :math:`\outputRV`, using the same multivariate orthonormal basis for
 all the marginals.
 Thus, the method is detailed here for a scalar output :math:`Y` and
-:math:`g: \Rset^{n_X} \rightarrow \Rset`.
+:math:`\model: \Rset^{\inputDim} \rightarrow \Rset`.
 
[…]

[…]
