### 2026-04-03 3558c2f
CI/abi3: Bumped swig
