+++ .circleci/run_docker_linux.sh
@@ -28,7 +28,6 @@ cmake -DCMAKE_INSTALL_PREFIX=${HOME}/.local \
       -DCMAKE_C_FLAGS="-Wall -Wextra -Wpedantic -Werror" \
       -DCMAKE_CXX_FLAGS="-Wall -Wextra -Wpedantic -Wshadow -Wsuggest-override -Werror -D_GLIBCXX_ASSERTIONS" \
       -DCMAKE_LINKER_TYPE=MOLD \
-      -DBLA_VENDOR=Generic \
       -DSWIG_COMPILE_FLAGS="-O1 -Wno-unused-parameter -Wno-shadow" -DSWIG_CXX_COMPILE_FLAGS="-Wno-suggest-override" \
       -DUSE_SPHINX=ON -DSPHINX_FLAGS="-W -T -j4" \
       ${source_dir}
