### 2026-04-08 66eb8ab
Improved Drawable
Now, it is possible to build a list of colors based on a list of scalars and a palette.
The rounding between scalars and 8 bit representations of color and alpha channels has been unified.
