### 2026-04-08 509454a
Improved Field drawing capabilities
Now one can draw a given marginal value of field in 3D, either through the draw(), drawMarginal() or draw3D() methods. The last one allows to fine-tune the rendering.
