+++ python/test/t_FORM_std.py
@@ -44,6 +44,10 @@ myAlgo = ot.FORM(myCobyla, myEvent)
 
 print("FORM=", myAlgo)
 
+# Check the starting point
+point = myAlgo.getPhysicalStartingPoint()
+myAlgo.setPhysicalStartingPoint(point)
+
 # Perform the simulation
 myAlgo.run()
 
