### 2026-04-14 6c39e6d
Sphinx: Misc fixes
