### 2026-04-14 1e82309
Sphinx: Misc fixes
