+++ python/doc/examples/numerical_methods/optimization/README.txt
@@ -1,4 +1,4 @@
-.. _Optimization:
+.. _optimization_examples:
 
 Optimization
 ------------
