+++ CMakeLists.txt
@@ -523,7 +523,7 @@ if (BUILD_PYTHON)
       if (SPHINX_EXECUTABLE AND (DVISVGM_EXECUTABLE OR DVIPNG_EXECUTABLE))
         find_python_module (sphinx 1.8)
         find_python_module (numpydoc 0.9)
-        find_python_module (sphinx_gallery 0.7)
+        find_python_module (sphinx_gallery 0.17)
         find_python_module (sphinx_copybutton)
         find_python_module (sphinxcontrib.jquery)
       endif ()
