### 2026-04-14 345518a
Sphinx: Misc fixes
