### 2026-04-15 cdbecbb
Sphinx: Drop jquery
Closes #2480
