### 2026-04-15 ccb79fc
Sphinx: Drop jquery
Closes #2480
