+++ CMakeLists.txt
@@ -525,11 +525,10 @@ if (BUILD_PYTHON)
         find_python_module (numpydoc 0.9)
         find_python_module (sphinx_gallery 0.17)
         find_python_module (sphinx_copybutton)
-        find_python_module (sphinxcontrib.jquery)
       endif ()
-      if (NOT NUMPYDOC_FOUND OR NOT SPHINX_GALLERY_FOUND OR NOT SPHINX_COPYBUTTON_FOUND OR NOT SPHINXCONTRIB.JQUERY_FOUND OR NOT MATPLOTLIB_FOUND OR NOT LIBXML2_FOUND)
+      if (NOT NUMPYDOC_FOUND OR NOT SPHINX_GALLERY_FOUND OR NOT SPHINX_COPYBUTTON_FOUND OR NOT MATPLOTLIB_FOUND OR NOT LIBXML2_FOUND)
         set (SPHINX_FOUND FALSE)
-        message (STATUS "Disabling documentation (need dvisvgm/numpydoc/sphinx-gallery/sphinx-copybutton/sphinxcontrib-jquery/matplotlib/libxml2)")
+        message (STATUS "Disabling documentation (need dvisvgm/numpydoc/sphinx-gallery/sphinx-copybutton/matplotlib/libxml2)")
       endif ()
     endif ()
 
