### 2026-04-15 c8f2352
Sphinx: Drop jquery
Closes #2480
