### 2026-04-15 34b78fa
CMake: Add python layout option
