+++ CMakeLists.txt
@@ -532,7 +532,9 @@ if (BUILD_PYTHON)
       endif ()
     endif ()
 
-    if (WIN32)
+    option (ENABLE_MSVC_PYTHON_LAYOUT "Force msvc python site-packages layout on win32 targets" ON)
+    mark_as_advanced (ENABLE_MSVC_PYTHON_LAYOUT)
+    if (WIN32 AND ENABLE_MSVC_PYTHON_LAYOUT)
       set (OPENTURNS_PYTHON_MODULE_PATH Lib/site-packages CACHE STRING "site-packages dir")
     else ()
       set (OPENTURNS_PYTHON_MODULE_PATH ${CMAKE_INSTALL_LIBDIR}/python${Python_VERSION_MAJOR}.${Python_VERSION_MINOR}/site-packages CACHE STRING "site-packages dir")
