### 2026-04-16 c9e4317
CMake: Mark variables as advanced
