### 2026-04-16 96de393
SymbolicEvaluation: Add doc
Closes #2716
