### 2026-04-16 5d47f45
Improved LeastSquaresExpansion
Now the basis size is adapted from a given list of active functions.
