### 2026-04-16 ca909cc
UserDefined: Rename to FiniteDiscreteDistribution
Closes #2519
