+++ lib/src/Uncertainty/Distribution/CMakeLists.txt
@@ -171,7 +171,7 @@ ot_add_source_file (UniformOverMesh.cxx)
 ot_add_source_file (UniformOrderStatistics.cxx)
 ot_add_source_file (UniformFactory.cxx)
 ot_add_source_file (UniformMuSigma.cxx)
-ot_add_source_file (UserDefined.cxx)
+ot_add_source_file (FiniteDiscreteDistribution.cxx)
 ot_add_source_file (UserDefinedFactory.cxx)
 ot_add_source_file (VonMises.cxx)
 ot_add_source_file (VonMisesFactory.cxx)
@@ -349,7 +349,7 @@ ot_install_header_file (UniformOverMesh.hxx)
 ot_install_header_file (UniformOrderStatistics.hxx)
 ot_install_header_file (UniformMuSigma.hxx)
 ot_install_header_file (UserDefinedFactory.hxx)
-ot_install_header_file (UserDefined.hxx)
+ot_install_header_file (FiniteDiscreteDistribution.hxx)
 ot_install_header_file (VonMises.hxx)
 ot_install_header_file (VonMisesFactory.hxx)
 ot_install_header_file (WeibullMax.hxx)
