### 2026-04-16 a1766b1
UserDefined: Rename to FiniteDiscreteDistribution
Closes #2519
