+++ python/doc/examples/data_analysis/sample_analysis/plot_visualize_empirical_cdf.py
@@ -19,11 +19,11 @@ normal = ot.Normal(1)
 sample = normal.getSample(size)
 
 # %%
-# We draw the empirical CDF based on the :class:`~openturns.UserDefined` distribution.
+# We draw the empirical CDF based on the :class:`~openturns.FiniteDiscreteDistribution` distribution.
 # By default, the `drawCDF` method requires no input argument.
 
 # %%
-distribution = ot.UserDefined(sample)
+distribution = ot.FiniteDiscreteDistribution(sample)
 graph = distribution.drawCDF()
 view = otv.View(graph)
 
@@ -34,7 +34,7 @@ view = otv.View(graph)
 # %%
 xmin = sample.getMin()[0] - 2.0
 xmax = sample.getMax()[0] + 2.0
-graph = ot.UserDefined(sample).drawCDF(xmin, xmax)
+graph = ot.FiniteDiscreteDistribution(sample).drawCDF(xmin, xmax)
 view = otv.View(graph)
 
 # %%
