### 2026-04-16 0932e44
UserDefined: Rename to FiniteDiscreteDistribution
Closes #2519
