+++ python/doc/examples/design_of_experiments/plot_design_of_experiment_continuous_discrete.py
@@ -13,14 +13,14 @@ import openturns.viewer as otv
 # %%
 # To create the first marginal of the distribution, we select a univariate discrete distribution.
 # Some of them, like the `Bernoulli` or `Geometric` distributions, are implemented in the library as classes.
-# In this example however, we pick the `UserDefined` distribution that assigns equal weights to the values -2, -1, 1 and 2.
+# In this example however, we pick the `FiniteDiscreteDistribution` distribution that assigns equal weights to the values -2, -1, 1 and 2.
 
 # %%
 sample = ot.Sample([[-2.0], [-1.0], [1.0], [2.0]])
 sample
 
 # %%
-X0 = ot.UserDefined(sample)
+X0 = ot.FiniteDiscreteDistribution(sample)
 
 # %%
 # For the second marginal, we pick a Gaussian distribution.
