+++ python/test/t_UserDefinedFactory_std.py
@@ -17,8 +17,8 @@ estimatedDistribution = factory.build()
 print("Default distribution=", estimatedDistribution)
 estimatedUserDefined = factory.buildAsUserDefined(sample)
 print("Sample                =", sample)
-print("Estimated UserDefined=", estimatedUserDefined)
+print("Estimated FiniteDiscreteDistribution=", estimatedUserDefined)
 estimatedUserDefined = factory.buildAsUserDefined()
-print("Default UserDefined=", estimatedUserDefined)
+print("Default FiniteDiscreteDistribution=", estimatedUserDefined)
 estimatedDistribution = factory.build([[10]] * 2000)
 print("Estimated distribution=", repr(estimatedDistribution))
