### 2026-04-16 05094ef
RandomMixture: Rename to LinearCombinationDistribution
Closes #3072
