### 2026-04-16 67e2ee4
RandomMixture: Rename to LinearCombinationDistribution
Closes #3072
