+++ python/test/t_LinearCombinationDistribution_simplification.py
@@ -3,23 +3,23 @@
 import openturns as ot
 
 ot.TESTPREAMBLE()
-ot.ResourceMap.SetAsUnsignedInteger("RandomMixture-DefaultMaxSize", 4000000)
+ot.ResourceMap.SetAsUnsignedInteger("LinearCombinationDistribution-DefaultMaxSize", 4000000)
 ot.TTY.ShowColors(False)
 
 # Test fusion of Dirac with no other atoms: should be a unique Dirac(5.8)
-d = ot.RandomMixture(
+d = ot.LinearCombinationDistribution(
     [ot.Dirac(1.0), ot.Dirac(2.0), ot.Dirac(3.0)], [0.5, 0.6, 0.7], 2.0
 )
 print("d=", d)
 # Test fusion of Dirac with other atoms: the Dirac should be merged into
 # the constant: 5 + Exponential(lambda=1,gamma=0)
-d = ot.RandomMixture([ot.Dirac(1.0), ot.Dirac(2.0), ot.Exponential()], 2.0)
+d = ot.LinearCombinationDistribution([ot.Dirac(1.0), ot.Dirac(2.0), ot.Exponential()], 2.0)
 print("d=", d)
-# Test flatten RandomMixture atoms: the RandomMixture should have 4 atoms.
-d = ot.RandomMixture(
+# Test flatten LinearCombinationDistribution atoms: the LinearCombinationDistribution should have 4 atoms.
+d = ot.LinearCombinationDistribution(
     [
         ot.Gumbel(),
-        ot.RandomMixture([ot.Logistic(), ot.WeibullMin()], [0.5, 1.5], 3.0),
+        ot.LinearCombinationDistribution([ot.Logistic(), ot.WeibullMin()], [0.5, 1.5], 3.0),
         ot.Frechet(1.0, 4.0),
     ],
     [2.0, 3.0, 4.0],
@@ -27,12 +27,12 @@ d = ot.RandomMixture(
 )
 print("d=", d)
 # Test merge of Normal atoms:
-d = ot.RandomMixture(
+d = ot.LinearCombinationDistribution(
     [ot.Normal(1.0, 8.0), ot.Logistic(), ot.Normal(2.0, 1.0)], [0.5, 2.5, 3.0], 2.0
 )
 print("d=", d)
 # Test merge of Exponential, Gamma and ChiSquare atoms
-d = ot.RandomMixture(
+d = ot.LinearCombinationDistribution(
     [
         ot.Exponential(1.0, 1.0),
         ot.Exponential(1.5, -1.0),
@@ -46,7 +46,7 @@ d = ot.RandomMixture(
 )
 print("d=", d)
 # Test merge of Uniform atoms
-d = ot.RandomMixture(
+d = ot.LinearCombinationDistribution(
     [
         ot.Uniform(0.0, 1.0),
         ot.Uniform(0.0, 1.0),
@@ -60,8 +60,8 @@ d = ot.RandomMixture(
 print("d=", d)
 # Test merge of Bernoulli and Binomial atoms
 # Deactivate the fusion of discrete atoms
-ot.ResourceMap.SetAsUnsignedInteger("RandomMixture-MaximumSupportSize", 0)
-d = ot.RandomMixture(
+ot.ResourceMap.SetAsUnsignedInteger("LinearCombinationDistribution-MaximumSupportSize", 0)
+d = ot.LinearCombinationDistribution(
     [
         ot.Bernoulli(0.5),
         ot.Bernoulli(0.5),
@@ -76,8 +76,8 @@ d = ot.RandomMixture(
 print("d=", d)
 # Test merge of Poisson atoms
 # Deactivate the fusion of discrete atoms
[…]
