+++ python/test/t_LinearCombinationDistribution_std.py
@@ -6,10 +6,10 @@ from math import sqrt, pi, exp, log
 
 ot.TESTPREAMBLE()
 
-ot.ResourceMap.SetAsUnsignedInteger("RandomMixture-DefaultMaxSize", 4000000)
+ot.ResourceMap.SetAsUnsignedInteger("LinearCombinationDistribution-DefaultMaxSize", 4000000)
 # Deactivate the simplification mechanism as we want to test the Poisson formula
 # based algorithm here
-ot.ResourceMap.SetAsBool("RandomMixture-SimplifyAtoms", False)
+ot.ResourceMap.SetAsBool("LinearCombinationDistribution-SimplifyAtoms", False)
 
 # Create a collection of test-cases and the associated references
 numberOfTests = 3
@@ -25,7 +25,7 @@ print("testCases=", testCases)
 print("references=", references)
 for testIndex in range(len(testCases)):
     # Instantiate one distribution object
-    distribution = ot.RandomMixture(testCases[testIndex])
+    distribution = ot.LinearCombinationDistribution(testCases[testIndex])
     distribution.setBlockMin(5)
     distribution.setBlockMax(20)
     distributionReference = references[testIndex]
@@ -169,12 +169,12 @@ coll.add(ot.Uniform(2.0, 4.0))
 weights.add(2.0)
 coll.add(ot.Exponential(2.0, -3.0))
 weights.add(1.5)
-rm = ot.RandomMixture(coll, weights)
+rm = ot.LinearCombinationDistribution(coll, weights)
 coll.add(rm)
 weights.add(-2.5)
 coll.add(ot.Gamma(3.0, 4.0, -2.0))
 weights.add(2.5)
-distribution = ot.RandomMixture(coll, weights)
+distribution = ot.LinearCombinationDistribution(coll, weights)
 print("distribution=", repr(distribution))
 print("distribution=", distribution)
 mu = distribution.getMean()[0]
@@ -196,7 +196,7 @@ result, norms = distribution.project(collFactories)
 print("projections=", result)
 print("norms=", norms)
 # ------------------------------ Multivariate tests ------------------------------#
-# 2D RandomMixture
+# 2D LinearCombinationDistribution
 collection = [ot.Normal(0.0, 1.0)] * 3
 
 weightMatrix = ot.Matrix(2, 3)
@@ -207,8 +207,8 @@ weightMatrix[1, 0] = 1.0
 weightMatrix[1, 1] = 1.0
 weightMatrix[1, 2] = -3.0
 
-# Build the RandomMixture
-distribution2D = ot.RandomMixture(collection, weightMatrix)
+# Build the LinearCombinationDistribution
+distribution2D = ot.LinearCombinationDistribution(collection, weightMatrix)
 print("distribution = ", distribution2D)
 print("range = ", distribution2D.getRange())
 print("mean = ", distribution2D.getMean())
@@ -250,8 +250,8 @@ for index in range(grid.getSize()):
 
 # 2D test, but too much CPU consuming
 collUniforme = [ot.Uniform(0, 1)] * 3
-# Build the RandomMixture
-dist_2D = ot.RandomMixture(collUniforme, weightMatrix)
+# Build the LinearCombinationDistribution
+dist_2D = ot.LinearCombinationDistribution(collUniforme, weightMatrix)
 dist_2D.setBlockMin(3)
 dist_2D.setBlockMax(8)
 
[…]
