### 2026-04-16 2225d66
RandomMixture: Rename to LinearCombinationDistribution
Closes #3072
