+++ lib/src/Uncertainty/Distribution/CMakeLists.txt
@@ -143,7 +143,7 @@ ot_add_source_file (PoissonFactory.cxx)
 ot_add_source_file (PosteriorDistribution.cxx)
 ot_add_source_file (ProductDistribution.cxx)
 ot_add_source_file (QuantileMatchingFactory.cxx)
-ot_add_source_file (RandomMixture.cxx)
+ot_add_source_file (LinearCombinationDistribution.cxx)
 ot_add_source_file (Rayleigh.cxx)
 ot_add_source_file (RayleighFactory.cxx)
 ot_add_source_file (Rice.cxx)
@@ -320,7 +320,7 @@ ot_install_header_file (Poisson.hxx)
 ot_install_header_file (PosteriorDistribution.hxx)
 ot_install_header_file (ProductDistribution.hxx)
 ot_install_header_file (QuantileMatchingFactory.hxx)
-ot_install_header_file (RandomMixture.hxx)
+ot_install_header_file (LinearCombinationDistribution.hxx)
 ot_install_header_file (RayleighFactory.hxx)
 ot_install_header_file (Rayleigh.hxx)
 ot_install_header_file (Rice.hxx)
