### 2026-04-16 c7535c4
CI: Use build dir option
