+++ .circleci/run_docker_mingw.sh
@@ -11,8 +11,6 @@ else
   source_dir=`pwd`  # circleci
 fi
 
-mkdir build && cd build
-
 ARCH=x86_64
 MINGW_PREFIX=/usr/${ARCH}-w64-mingw32
 PYMAJMIN=310
@@ -27,7 +25,8 @@ ${ARCH}-w64-mingw32-cmake \
   -DPython_LIBRARY=${MINGW_PREFIX}/lib/libpython${PYMAJMIN}.dll.a \
   -DPython_EXECUTABLE=/usr/bin/${ARCH}-w64-mingw32-python${PYMAJMIN}-bin \
   -DCMAKE_UNITY_BUILD=ON -DCMAKE_UNITY_BUILD_BATCH_SIZE=32 \
-  ${source_dir}
+  -B build ${source_dir}
+cd build
 make install
 ${ARCH}-w64-mingw32-strip --strip-unneeded ${PREFIX}/bin/*.dll ${PREFIX}/Lib/site-packages/openturns/*.pyd
 echo lib/test ${PREFIX}/Lib/site-packages/openturns | xargs -n 1 cp ${PREFIX}/bin/*.dll
