### 2026-04-16 073383f
Improved TruncatedDistribution
Now, the simplification algorithm takes into account the KernelMixture case.
It improves things only if the dimension is greater than or equal to 3.
It fixes #3162.
