+++ python/src/viewer.py
@@ -17,7 +17,7 @@ import openturns as ot
 import numpy as np
 import matplotlib
 import matplotlib.pyplot as plt
-import matplotlib.colors as cls
+import matplotlib.colors as colors
 import warnings
 import io
 
@@ -29,7 +29,7 @@ except ImportError:
 __all__ = ["View", "PlotDesign"]
 
 
-class RankNormalize(cls.Normalize):
+class RankNormalize(colors.Normalize):
     """
     Color distribution normalization class based on rank and not value.
 
@@ -52,8 +52,8 @@ class RankNormalize(cls.Normalize):
                     1.0 + above_threshold / active,
                     len(self._levels),
                 )
-        if hasattr(cls.Normalize, "_changed"):
-            cls.Normalize._changed(self)
+        if hasattr(colors.Normalize, "_changed"):
+            colors.Normalize._changed(self)
 
     def __init__(self, levels, vmin=None, vmax=None, clip=False):
         """
@@ -86,7 +86,7 @@ class RankNormalize(cls.Normalize):
         if self._levels is not None and self._ranks is None:
             self._ranks = np.linspace(0.0, 1.0, len(self._levels))
         return (
-            1.0
+            np.ones_like(value)
             if self._levels is None
             else np.ma.masked_array(np.interp(value, self._levels, self._ranks))
         )
@@ -417,16 +417,16 @@ class View:
             text_kw = dict(text_kw_default)
 
             # use drawable order over default patch/line/text order
-            plot_kw["zorder"] = zorder
-            bar_kw["zorder"] = zorder
-            polygon_kw["zorder"] = zorder
-            polygoncollection_kw["zorder"] = zorder
-            contour_kw["zorder"] = zorder
-            step_kw["zorder"] = zorder
+            plot_kw.setdefault("zorder", zorder)
+            bar_kw.setdefault("zorder", zorder)
+            polygon_kw.setdefault("zorder", zorder)
+            polygoncollection_kw.setdefault("zorder", zorder)
+            contour_kw.setdefault("zorder", zorder)
+            step_kw.setdefault("zorder", zorder)
             if matplotlib_version >= Version("3.3"):
-                clabel_kw["zorder"] = zorder
-            scatter_kw["zorder"] = zorder
-            text_kw["zorder"] = zorder
+                clabel_kw.setdefault("zorder", zorder)
+            scatter_kw.setdefault("zorder", zorder)
+            text_kw.setdefault("zorder", zorder)
             zorder += 1 / (len(drawables) + 1)
 
             # set color
@@ -512,9 +512,9 @@ class View:
                     bar_kw["edgecolor"] = bar_kw["color"]
 
                 xi = drawable.getOrigin()
-                for i in range(x.getSize()):
+                for i in range(len(x)):
                     # label only the first bar to avoid getting several items
-                    if (i == 1) and ("label" in bar_kw):
+                    if i > 0 and "label" in bar_kw:
                         bar_kw.pop("label")
                     self._ax[0].bar(xi, height=y[i], width=x[i], align="edge", **bar_kw)
                     xi += x[i]
@@ -573,12 +573,13 @@ class View:
                     )
                 poly = matplotlib.patches.Polygon(data, **polygon_kw)
                 self._ax[0].add_patch(poly)
-                legend_handles.append(poly)
-                legend_labels.append(drawable.getLegend())
+                if len(drawable.getLegend()) > 0:
+                    legend_handles.append(poly)
+                    legend_labels.append(drawable.getLegend())
 
             elif drawableKind == "PolygonArray":
-                polygonsNumber = drawable.getPalette().getSize()
-                verticesNumber = drawable.getData().getSize() // polygonsNumber
[…]
