### 2026-04-17 68e61e5
Doc: Update install section
