### 2026-04-17 2d62372
Doc: Update install section
