### 2026-04-17 e360349
UserDefinedFactory: Rename to FiniteDiscDistFact
