### 2026-04-17 9507f24
UserDefinedFactory: Rename to FiniteDiscDistFact
