+++ python/test/t_FiniteDiscreteDistributionFactory_std.py
@@ -9,16 +9,16 @@ sample[0] = ot.Point((1.0, 1.5))
 sample[1] = ot.Point((2.0, 2.5))
 sample[2] = ot.Point((3.0, 3.5))
 
-factory = ot.UserDefinedFactory()
+factory = ot.FiniteDiscreteDistributionFactory()
 estimatedDistribution = factory.build(sample)
 print("sample=", repr(sample))
 print("Estimated distribution=", repr(estimatedDistribution))
 estimatedDistribution = factory.build()
 print("Default distribution=", estimatedDistribution)
-estimatedUserDefined = factory.buildAsUserDefined(sample)
+estimatedUserDefined = factory.buildAsFiniteDiscreteDistribution(sample)
 print("Sample                =", sample)
 print("Estimated FiniteDiscreteDistribution=", estimatedUserDefined)
-estimatedUserDefined = factory.buildAsUserDefined()
+estimatedUserDefined = factory.buildAsFiniteDiscreteDistribution()
 print("Default FiniteDiscreteDistribution=", estimatedUserDefined)
 estimatedDistribution = factory.build([[10]] * 2000)
 print("Estimated distribution=", repr(estimatedDistribution))
