### 2026-04-17 c4b460f
UserDefinedFactory: Rename to FiniteDiscDistFact
