+++ lib/src/Uncertainty/Distribution/CMakeLists.txt
@@ -172,7 +172,7 @@ ot_add_source_file (UniformOrderStatistics.cxx)
 ot_add_source_file (UniformFactory.cxx)
 ot_add_source_file (UniformMuSigma.cxx)
 ot_add_source_file (FiniteDiscreteDistribution.cxx)
-ot_add_source_file (UserDefinedFactory.cxx)
+ot_add_source_file (FiniteDiscreteDistributionFactory.cxx)
 ot_add_source_file (VonMises.cxx)
 ot_add_source_file (VonMisesFactory.cxx)
 ot_add_source_file (WeibullMax.cxx)
@@ -348,7 +348,7 @@ ot_install_header_file (Uniform.hxx)
 ot_install_header_file (UniformOverMesh.hxx)
 ot_install_header_file (UniformOrderStatistics.hxx)
 ot_install_header_file (UniformMuSigma.hxx)
-ot_install_header_file (UserDefinedFactory.hxx)
+ot_install_header_file (FiniteDiscreteDistributionFactory.hxx)
 ot_install_header_file (FiniteDiscreteDistribution.hxx)
 ot_install_header_file (VonMises.hxx)
 ot_install_header_file (VonMisesFactory.hxx)
