### 2026-04-17 9776935
MixedHistogramUserDefined: Rename to xxxFiniteDiscrete
