+++ lib/src/Uncertainty/Distribution/CMakeLists.txt
@@ -116,7 +116,7 @@ ot_add_source_file (MeixnerDistribution.cxx)
 ot_add_source_file (MeixnerDistributionFactory.cxx)
 ot_add_source_file (MethodOfMomentsFactory.cxx)
 ot_add_source_file (MinCopula.cxx)
-ot_add_source_file (MixedHistogramUserDefined.cxx)
+ot_add_source_file (MixedHistogramFiniteDiscreteDistribution.cxx)
 ot_add_source_file (Mixture.cxx)
 ot_add_source_file (Multinomial.cxx)
 ot_add_source_file (MultinomialFactory.cxx)
@@ -292,7 +292,7 @@ ot_install_header_file (MeixnerDistribution.hxx)
 ot_install_header_file (MeixnerDistributionFactory.hxx)
 ot_install_header_file (MethodOfMomentsFactory.hxx)
 ot_install_header_file (MinCopula.hxx)
-ot_install_header_file (MixedHistogramUserDefined.hxx)
+ot_install_header_file (MixedHistogramFiniteDiscreteDistribution.hxx)
 ot_install_header_file (Mixture.hxx)
 ot_install_header_file (MultinomialFactory.hxx)
 ot_install_header_file (Multinomial.hxx)
