+++ python/doc/theory/numerical_methods/distribution_realization.rst
@@ -29,7 +29,7 @@ Several classical techniques exist:
       normal distribution is numerically invertible over :math:`[A,B]`
       with :math:`A\simeq -17.3` and :math:`B\simeq 8.5`.
 
--  The rejection method: suppose that one want to sample a random
+-  The rejection method: suppose that one wants to sample a random
    variable :math:`X` with a continuous distribution with PDF
    :math:`p_X` and that we know how to sample a random variable
    :math:`Y` with a continuous distribution with PDF :math:`p_Y`. We
@@ -64,7 +64,7 @@ Several classical techniques exist:
    -  The test :math:`u\leq \rho(u)` can be replaced by an equivalent
       one but much more computationally efficient.
 
--  The transformation method: suppose that one want to sample a random
+-  The transformation method: suppose that one wants to sample a random
    variable that is the image by a simple transformation of another
    random variable (or random vector) that can easily be sampled. It is
    easy to sample this last random variable (or vector) and then
