### 2026-04-20 eb91e9e
SWIG: Move MeshDomain to algo
