### 2026-04-20 5658793
SWIG: Move P1Lagrange* to algo
