### 2026-04-20 4f58d30
SWIG: Split func module
