### 2026-04-20 eba3db4
SWIG: Split model_copula
