+++ python/src/CMakeLists.txt
@@ -83,43 +83,6 @@ endmacro ()
 
 install (DIRECTORY openturns DESTINATION ${OPENTURNS_INCLUDE_PATH})
 
-ot_add_python_module(common common_module.i
-                      OTtypes.i
-                      OTexceptions.i
-                      TBB.i TBB_doc.i
-                      Object.i Object_doc.i
-                      PersistentObject.i PersistentObject_doc.i
-                      Study.i Study_doc.i
-                      StorageManager.i StorageManager_doc.i
-                      XMLStorageManager.i XMLStorageManager_doc.i
-                      XMLH5StorageManager.i XMLH5StorageManager_doc.i
-                      Log.i Log_doc.i
-                      TTY.i TTY_doc.i
-                      Path.i Path_doc.i
-                      Catalog.i
-                      ResourceMap.i ResourceMap_doc.i
-                      ComparisonOperatorImplementation.i
-                      ComparisonOperator.i ComparisonOperator_doc.i
-                      Equal.i Equal_doc.i
-                      Greater.i Greater_doc.i
-                      GreaterOrEqual.i GreaterOrEqual_doc.i
-                      Less.i Less_doc.i
-                      LessOrEqual.i LessOrEqual_doc.i
-                      Cache.i
-                      Pointer.i
-                      InterfaceObject.i InterfaceObject_doc.i
-                      TypedInterfaceObject.i
-                      TypedCollectionInterfaceObject.i
-                      BaseCommonTemplateDefs.i
-                   )
-
-if (LibXml2_FOUND)
-  set_property (TARGET common APPEND PROPERTY SWIG_COMPILE_DEFINITIONS OPENTURNS_HAVE_LIBXML2=1)
-endif ()
-
-if (HDF5_FOUND)
-  set_property (TARGET common APPEND PROPERTY SWIG_COMPILE_DEFINITIONS OPENTURNS_HAVE_HDF5=1)
-endif ()
 
 ot_add_python_module(typ typ_module.i
                       Collection.i Collection_doc.i
@@ -153,109 +116,6 @@ ot_add_python_module(typ typ_module.i
                       HistoryStrategy.i HistoryStrategy_doc.i
                       BaseTypCollection.i
                       BaseTypTemplateDefs.i
-                   )
-
-ot_add_python_module(statistics statistics_module.i
-                      Compact.i Compact_doc.i
-                      Full.i Full_doc.i
-                      Last.i Last_doc.i
-                      Null.i Null_doc.i
-                      TestResult.i TestResult_doc.i
-                      RandomGenerator.i RandomGenerator_doc.i
-                      RandomGeneratorState.i RandomGeneratorState_doc.i
-                      LowDiscrepancySequenceImplementation.i LowDiscrepancySequenceImplementation_doc.i
-                      LowDiscrepancySequence.i LowDiscrepancySequence_doc.i
-                      SobolSequence.i SobolSequence_doc.i
-                      FaureSequence.i FaureSequence_doc.i
-                      HaltonSequence.i HaltonSequence_doc.i
-                      HaselgroveSequence.i HaselgroveSequence_doc.i
-                      ReverseHaltonSequence.i ReverseHaltonSequence_doc.i
-                      SphericalModel.i SphericalModel_doc.i
-                      FilteringWindowsImplementation.i FilteringWindowsImplementation_doc.i
-                      FilteringWindows.i FilteringWindows_doc.i
-                      Hamming.i Hamming_doc.i
-                      Hann.i Hann_doc.i
-                      FFTImplementation.i FFTImplementation_doc.i
-                      FFT.i FFT_doc.i
-                      SpectralModelImplementation.i SpectralModelImplementation_doc.i
[…]
