### 2026-04-21 31b3304
SWIG: Balance dist_bundle
