### 2026-04-22 86cd241
CMake: Add USE_GPL_LIBS option
Closes #3170
