+++ CMakeLists.txt
@@ -10,10 +10,11 @@ set (CMAKE_TRY_COMPILE_CONFIGURATION ${CMAKE_BUILD_TYPE})
 project (openturns)
 
 option (USE_EXTERNAL_LIBS            "Use external libs"                                                     ON)
-mark_as_advanced (USE_EXTERNAL_LIBS)
+option (USE_GPL_LIBS                 "Allow GPL-licensed libs (OT binary would be GPL instead of LGPL)"      OFF)
+mark_as_advanced (USE_EXTERNAL_LIBS USE_GPL_LIBS)
 option (USE_CUBA                     "Use Cuba library"                                                      ${USE_EXTERNAL_LIBS})
 option (USE_TBB                      "Use Intel Threading Building Blocks library for multithreading"        ${USE_EXTERNAL_LIBS})
-option (USE_HMAT                     "Use HMat library (enforces GPL license)"                               OFF)
+option (USE_HMAT                     "Use HMat library (enforces GPL license)"                               ${USE_GPL_LIBS})
 option (USE_MUPARSER                 "Use muParser library"                                                  ${USE_EXTERNAL_LIBS})
 option (USE_EXPRTK                   "Use ExprTK library"                                                    ON)
 option (USE_LIBXML2                  "Use LibXml2 for XML support"                                           ${USE_EXTERNAL_LIBS})
@@ -265,7 +266,7 @@ if (USE_CERES)
     message (STATUS "Ceres was built with miniglog and will likely leak logs to stderr (prefer glog), disabling.")
     set (Ceres_FOUND FALSE)
   endif ()
-  if (Ceres_FOUND AND "SuiteSparse" IN_LIST CERES_COMPILED_COMPONENTS)
+  if (Ceres_FOUND AND "SuiteSparse" IN_LIST CERES_COMPILED_COMPONENTS AND NOT USE_GPL_LIBS)
     message (STATUS "Ceres was built with SuiteSparse GPL components, disabling.")
     set (Ceres_FOUND FALSE)
   endif ()
