### 2026-04-24 b1d84cb
Doc: Explain ref/any usage
