+++ python/doc/examples/stochastic_processes/plot_field_manipulation.py
@@ -18,6 +18,7 @@ Draw a field
 # %%
 import openturns as ot
 import openturns.viewer as otv
+
 # sphinx_gallery_thumbnail_number = 2
 
 # %%
@@ -36,20 +37,20 @@ field = normal_proc.getRealization()
 # %%
 # We first draw the field with arrows fixed on each vertex of the mesh.
 g_1 = field.draw()
-g_1.setTitle(r'A field on 2-d mesh and 2-d values')
-g_1.setXTitle(r'$x_0$')
-g_1.setYTitle(r'$x_1$')
+g_1.setTitle(r"A field on 2-d mesh and 2-d values")
+g_1.setXTitle(r"$x_0$")
+g_1.setYTitle(r"$x_1$")
 
 # %%
 # In order to change the arrows, we can modify the :class:`~openturns.ResourceMap` keys associated to the
 # class :class:`~openturns.Field`.
-ot.ResourceMap.SetAsScalar('Field-ArrowRatio', 0.05)
-ot.ResourceMap.SetAsScalar('Field-ArrowScaling', 0.8)
+ot.ResourceMap.SetAsScalar("Field-ArrowRatio", 0.05)
+ot.ResourceMap.SetAsScalar("Field-ArrowScaling", 0.8)
 g_2 = field.draw()
 
-g_2.setTitle('The same field with other arrows parameters')
-g_2.setXTitle(r'$x_0$')
-g_2.setYTitle(r'$x_1$')
+g_2.setTitle("The same field with other arrows parameters")
+g_2.setXTitle(r"$x_0$")
+g_2.setYTitle(r"$x_1$")
 
 # %%
 # We can compare both figures.
@@ -61,14 +62,14 @@ view = otv.View(grid)
 # %%
 # We can also use the output values to deform the initial mesh and draw the deformed mesh.
 # First, we recover the default values of the ResourceMap keys.
-ot.ResourceMap.SetAsScalar('Field-ArrowRatio', 0.01)
-ot.ResourceMap.SetAsScalar('Field-ArrowScaling', 1)
+ot.ResourceMap.SetAsScalar("Field-ArrowRatio", 0.01)
+ot.ResourceMap.SetAsScalar("Field-ArrowScaling", 1)
 deformed_field = field.asDeformedMesh()
 g_def = deformed_field.draw()
-g_def.setTitle('The same field as deformed mesh')
-g_def.setXTitle(r'$x_0$')
-g_def.setYTitle(r'$x_1$')
-g_def.setLegendPosition('')
+g_def.setTitle("The same field as deformed mesh")
+g_def.setXTitle(r"$x_0$")
+g_def.setYTitle(r"$x_1$")
+g_def.setLegendPosition("")
 
 # %%
 # We can compare both figures.
@@ -91,18 +92,18 @@ field = normal_proc.getRealization()
 # %%
 # We draw the field by filling each cell of the mesh with a color.
 g_1 = field.draw()
-g_1.setTitle('A field on 2-d mesh  and scalar values')
-g_1.setXTitle(r'$x_0$')
-g_1.setYTitle(r'$x_1$')
+g_1.setTitle("A field on 2-d mesh  and scalar values")
+g_1.setXTitle(r"$x_0$")
+g_1.setYTitle(r"$x_1$")
 
 
 # %%
 # We can also draw iso-value lines of the output values using interpolation.
 g_2 = field.drawMarginal(0, True)
-g_2.setTitle('A field on 2-d mesh and scalar values')
-g_2.setXTitle(r'$x_0$')
-g_2.setYTitle(r'$x_1$')
-g_2.setLegendPosition('')
+g_2.setTitle("A field on 2-d mesh and scalar values")
[…]
