+++ python/test/t_MixedHistogramFiniteDiscreteDistribution_std.py
@@ -5,9 +5,14 @@ import openturns.testing as ott
 
 ticksCollection = [[0.0, 1.0, 3.0, 6.0], [1.0, 2.0, 3.0]]
 
-kind = [ot.MixedHistogramFiniteDiscreteDistribution.CONTINUOUS, ot.MixedHistogramFiniteDiscreteDistribution.DISCRETE]
+kind = [
+    ot.MixedHistogramFiniteDiscreteDistribution.CONTINUOUS,
+    ot.MixedHistogramFiniteDiscreteDistribution.DISCRETE,
+]
 probabilityTable = [0.05, 0.1, 0.15, 0.2, 0.25, 0.3, 0.35, 0.4, 0.45]
-distribution = ot.MixedHistogramFiniteDiscreteDistribution(ticksCollection, kind, probabilityTable)
+distribution = ot.MixedHistogramFiniteDiscreteDistribution(
+    ticksCollection, kind, probabilityTable
+)
 distribution.setDescription(["A", "B"])
 print("Distribution ", repr(distribution))
 print("Distribution ", distribution)
@@ -80,7 +85,9 @@ print("probabilityTable=", probabilityTable)
 x = [3.0] * dim
 x[dim - 2] = 5.0
 bench = 0
-distribution = ot.MixedHistogramFiniteDiscreteDistribution(ticksCollection, kind, probabilityTable)
+distribution = ot.MixedHistogramFiniteDiscreteDistribution(
+    ticksCollection, kind, probabilityTable
+)
 print("pdf=%.6g" % distribution.computePDF(x))
 print("cdf=%.6g" % distribution.computeCDF(x))
 
