+++ python/src/usecases/flood_model.py
@@ -168,7 +168,9 @@ class FloodModel:
         formula += "var Zc := H + Zv;"
         formula += "var Zd := Zb + Hd;"
         formula += "S := Zc - Zd;"
-        formula += "var Cost_Flooding := if (S < 0, 0.2 - 0.8 * expm1(-1000 / S^4), 1.0);"
+        formula += (
+            "var Cost_Flooding := if (S < 0, 0.2 - 0.8 * expm1(-1000 / S^4), 1.0);"
+        )
         formula += "var Cost_Dyke := max(Hd, 8.0) / 20.0;"
         formula += "C := Cost_Flooding + Cost_Dyke;"
 
