+++ python/doc/pyplots/P1LagrangeInterpolation.py
@@ -29,10 +29,10 @@ graph_1 = input_field.drawMarginal(0)
 graph_1.setXTitle("$x$")
 graph_1.setYTitle("$y$")
 graph_1.setTitle(r"Function $x \rightarrow sin(\pi x)sin(\pi y)$ on the input mesh")
-graph_1.setLegendPosition('')
+graph_1.setLegendPosition("")
 
 graph_2 = output_field.drawMarginal(0)
-graph_2.setLegendPosition('')
+graph_2.setLegendPosition("")
 graph_2.setXTitle("$x$")
 graph_2.setYTitle("$y$")
 graph_2.setTitle(r"Function $x \rightarrow sin(\pi x)sin(\pi y)$ on the output mesh")

[…]
