+++ python/doc/examples/probabilistic_modeling/distributions/plot_compound_distribution.py
@@ -55,7 +55,7 @@ XgivenThetaDist = ot.Uniform()
 # We advice to change the discretization method that performs the integration.
 # By default, this method is the quadrature method *GaussProduct*. We prefer to use the *QMC*.
 # We use the :class:`~openturns.ResourceMap`.
-ot.ResourceMap.SetAsString('CompoundDistribution-ContinuousDiscretizationMethod', 'QMC')
+ot.ResourceMap.SetAsString("CompoundDistribution-ContinuousDiscretizationMethod", "QMC")
 
 # %%
 # Create the compound distribution of :math:`X`.
