+++ python/test/t_SubsetSampling_R-S.py
@@ -62,7 +62,9 @@ resultMC = myMC.getResult()
 PFMC = resultMC.getProbabilityEstimate()
 CVMC = resultMC.getCoefficientOfVariation()
 variance_PF_MC = resultMC.getVarianceEstimate()
-length90MC = resultMC.getProbabilityDistribution().computeBilateralConfidenceInterval(0.9)
+length90MC = resultMC.getProbabilityDistribution().computeBilateralConfidenceInterval(
+    0.9
+)
 N_MC = resultMC.getOuterSampling() * resultMC.getBlockSize()
 
 #
@@ -71,7 +73,9 @@ resultSS = mySS.getResult()
 PFSS = resultSS.getProbabilityEstimate()
 CVSS = resultSS.getCoefficientOfVariation()
 variance_PF_SS = resultSS.getVarianceEstimate()
-length90SS = resultSS.getProbabilityDistribution().computeBilateralConfidenceInterval(0.9)
+length90SS = resultSS.getProbabilityDistribution().computeBilateralConfidenceInterval(
+    0.9
+)
 N_SS = resultSS.getOuterSampling() * resultSS.getBlockSize()
 
 #
