+++ python/test/t_LinearCombinationDistribution_simplification.py
@@ -3,7 +3,9 @@
 import openturns as ot
 
 ot.TESTPREAMBLE()
-ot.ResourceMap.SetAsUnsignedInteger("LinearCombinationDistribution-DefaultMaxSize", 4000000)
+ot.ResourceMap.SetAsUnsignedInteger(
+    "LinearCombinationDistribution-DefaultMaxSize", 4000000
+)
 ot.TTY.ShowColors(False)
 
 # Test fusion of Dirac with no other atoms: should be a unique Dirac(5.8)
@@ -13,13 +15,17 @@ d = ot.LinearCombinationDistribution(
 print("d=", d)
 # Test fusion of Dirac with other atoms: the Dirac should be merged into
 # the constant: 5 + Exponential(lambda=1,gamma=0)
-d = ot.LinearCombinationDistribution([ot.Dirac(1.0), ot.Dirac(2.0), ot.Exponential()], 2.0)
+d = ot.LinearCombinationDistribution(
+    [ot.Dirac(1.0), ot.Dirac(2.0), ot.Exponential()], 2.0
+)
 print("d=", d)
 # Test flatten LinearCombinationDistribution atoms: the LinearCombinationDistribution should have 4 atoms.
 d = ot.LinearCombinationDistribution(
     [
         ot.Gumbel(),
-        ot.LinearCombinationDistribution([ot.Logistic(), ot.WeibullMin()], [0.5, 1.5], 3.0),
+        ot.LinearCombinationDistribution(
+            [ot.Logistic(), ot.WeibullMin()], [0.5, 1.5], 3.0
+        ),
         ot.Frechet(1.0, 4.0),
     ],
     [2.0, 3.0, 4.0],
@@ -60,7 +66,9 @@ d = ot.LinearCombinationDistribution(
 print("d=", d)
 # Test merge of Bernoulli and Binomial atoms
 # Deactivate the fusion of discrete atoms
-ot.ResourceMap.SetAsUnsignedInteger("LinearCombinationDistribution-MaximumSupportSize", 0)
+ot.ResourceMap.SetAsUnsignedInteger(
+    "LinearCombinationDistribution-MaximumSupportSize", 0
+)
 d = ot.LinearCombinationDistribution(
     [
         ot.Bernoulli(0.5),
@@ -76,7 +84,9 @@ d = ot.LinearCombinationDistribution(
 print("d=", d)
 # Test merge of Poisson atoms
 # Deactivate the fusion of discrete atoms
-ot.ResourceMap.SetAsUnsignedInteger("LinearCombinationDistribution-MaximumSupportSize", 0)
+ot.ResourceMap.SetAsUnsignedInteger(
+    "LinearCombinationDistribution-MaximumSupportSize", 0
+)
 d = ot.LinearCombinationDistribution(
     [
         ot.Poisson(3.0),
@@ -91,7 +101,9 @@ d = ot.LinearCombinationDistribution(
 print("d=", d)
 # Test merge of discrete and continuous atoms
 # Deactivate the fusion of discrete atoms
-ot.ResourceMap.SetAsUnsignedInteger("LinearCombinationDistribution-MaximumSupportSize", 0)
+ot.ResourceMap.SetAsUnsignedInteger(
+    "LinearCombinationDistribution-MaximumSupportSize", 0
+)
 # more continuous atoms
 d = ot.LinearCombinationDistribution(
     [ot.Logistic(), ot.Binomial(2, 0.5), ot.Uniform()], [1.0, 2.0, 3.0], 2.0
@@ -113,7 +125,9 @@ print("d=", d)
 ot.Log.Show(ot.Log.DBG)
 print("\n")
[…]
