+++ python/test/t_Waarts_system_series.py
@@ -172,7 +172,9 @@ ResultMC = myMC.getResult()
 PFMC = ResultMC.getProbabilityEstimate()
 CVMC = ResultMC.getCoefficientOfVariation()
 Variance_PF_MC = ResultMC.getVarianceEstimate()
-length90MC = ResultMC.getProbabilityDistribution().computeBilateralConfidenceInterval(0.9)
+length90MC = ResultMC.getProbabilityDistribution().computeBilateralConfidenceInterval(
+    0.9
+)
 
 #
 # Outputs

[…]
