### 2026-04-24 d39fe32
GaussianProcessFitter: Desambiguate ref
