+++ python/doc/theory/meta_modeling/gaussian_process_regression.rst
@@ -114,7 +114,7 @@ where :math:`\mat{C}_{ij} = C_{\vect{p}}(\vect{x}_i, \vect{x}_j)\in \cS_{\output
 
 The likelihood of the Gaussian process on the data set is defined by:
 
-.. _logLikelihoodGP:
+.. _logLikelihoodGPmodel:
 
 .. math::
     :label: logLikelihoodGPgen
