### 2026-04-24 2077d24
R script to compute LMG and PMVD indices
