+++ validation/src/Multicollinearity.R
@@ -1,0 +1,179 @@
+# This script computes variance-based importance measures (LMG and PMVD) in the context of linear regression
+# The test cases are borrowed from https://gitlab.com/LauraClouvel/toydata/
+
+library(relaimpo)
+library(sensitivity)
+library(AmesHousing)
+library(caret)
+library(mlbench)
+library(mvtnorm)
+
+#
+# Compute a measure for the given variables
+#
+# Parameters
+# - x: a data.frame containing the input variables
+# - y: a vector containing the output variable
+# - type: the measure type ("LMG" or "PMVD")
+# - package: the name of the package that will compute the measure ("relaimpo" or "sensitivity")
+#
+# Output
+#   A vector containing the measure for each input variable
+#
+compute_measure <- function(x, y, type, package)
+{
+  if (type=="LMG")
+  {
+    if (package=="relaimpo")
+    {
+      res <- calc.relimp(y, x, type="lmg")
+      return(res$lmg)
+    }
+    else if (package=="sensitivity")
+    {
+      res <- lmg(x, y)
+      return(res$lmg[,1])
+    }
+    else
+      stop("Unknown package")
+  }
+  else if (type=="PMVD")
+  {
+    if (package=="relaimpo")
+    {
+      res <- calc.relimp(y, x, type="pmvd")
+      return(res$pmvd)
+    }
+    else if (package=="sensitivity")
+    {
+      res <- pmvd(x, y)
+      return(res$pmvd[,1])
+    }
+    else
+      stop("Unknown package")
+  }
+  else
+    stop("Unknown type")
+}
+
+header <- function(text)
+{
+  cat(strrep("-", 50), "\n")
+  cat(text, "\n")
+  cat(strrep("-", 50), "\n")
+  cat("\n")
+}
+
+get_data1 <- function()
+{
+  header("Case #1 - Linear model: Y = X1^2 + X2 + X3")
+  n <- 100
+  x <- data.frame(X1 = runif(n, 0.5, 1.5), X2 = runif(n, 1.5, 4.5), X3 = runif(n, 4.5, 13.5))
+  y <- with(x, X1^2 + X2 + X3)
+  return(list(x=x, y=y, rep=100))
+}
+
+get_data2 <- function()
+{
+  header("Case #2 - Linear model: Y = X1^2 + X2 + X3 + X4")
+  n <- 100
+  x <- data.frame(X1 = runif(n, 0.5, 1.5), X2 = runif(n, 1.5, 4.5), X3 = runif(n, 4.5, 13.5))
+  x <- data.frame(x, X4 = x$X3 + rnorm(n))
+  y <- with(x, X1^2 + X2 + X3 + X4)
+  return(list(x=x, y=y, rep=100))
+}
+
+get_data3 <- function()
+{
+  header("Case #3 - Linear model: Y = X1 + eta")
+  n <- 100
+  xx <- rmvnorm(n, mean=c(0,0), sigma=matrix(c(1,0.9,0.9,1), ncol=2))
+  x <- data.frame(X1 = xx[,1], X2 = xx[,2])
+  y <- with(x, X1 + rnorm(n,0,0.1))
+  return(list(x=x, y=y, rep=100))
+}
+
+get_data4 <- function()
+{
+  header("Case #4 - Linear model: Y = X1 + 2X2 + eta")
+  n <- 100
+  xx <- rmvnorm(n, mean=c(0,0), sigma=matrix(c(1,0.9,0.9,1), ncol=2))
+  x <- data.frame(X1 = xx[,1], X2 = xx[,2])
+  y <- with(x, X1 + 2 * X2 + rnorm(n,0,0.1))
+  return(list(x=x, y=y, rep=100))
+}
+
+get_data5 <- function()
+{
[…]
