### 2026-04-27 9a39542
Doc: Update installation instructions
