### 2026-04-28 #10116 @Nikoleta-v3
Thank you for the submission, @mbaudin47. I will start inviting reviewers. I saw that you have suggested one reviewer; if there are any others you would recommend, please let me know.

Regarding the paper, there is one small typo in the AI usage disclosure: it should read that "the paper complies *with* the requirements."
