### 2026-04-28 #10116 @abhijit-c
Just to clarify, you wrote:

> The submission is titled: “biostats: Biostatistics and Clinical Data Analysis in R". You can find more details at the top of the GitHub issue for the submission. 🆙

but this is the submission for OpenTURNS' JOSS submission. Has there been a mistake somewhere?

If this is indeed intended for OpenTURNS, I'd be happy to review. I've used the library before and work in the space. If it's for `biostats`, I'd have to respectfully refuse; I don't know R!
