### 2026-04-28 #10116 @Nikoleta-v3
Hello @abhijit-c and @dannys4, thank you for your prompt responses! Apologies I copied and pasted that message from another submission I am handling 😅🙏🏻

The submission I’m asking you to review is “OpenTURNS for probabilistic modeling and uncertainty quantification.” 🚀
