+++ python/test/t_LineSampling_std.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
 
 import openturns as ot
-import openturns.experimental as otexp
 import openturns.testing as ott
 from openturns.usecases import cantilever_beam
 from openturns.usecases import stressed_beam
@@ -37,7 +36,7 @@ print(f"alpha={alpha}")
 ot.RandomGenerator.SetSeed(0)
 solver = ot.Brent(1e-3, 1e-3, 1e-3, 5)
 rootStrategy = ot.MediumSafe(solver)
-algo = otexp.LineSampling(event, alpha, rootStrategy)
+algo = ot.LineSampling(event, alpha, rootStrategy)
 algo.setMaximumCoefficientOfVariation(1e-3)
 algo.setMaximumOuterSampling(1000)
 algo.setStoreHistory(True)
@@ -86,7 +85,7 @@ print(f"alpha={alpha}")
 ot.RandomGenerator.SetSeed(0)
 solver = ot.Brent(1e-3, 1e-3, 1e-3, 5)
 rootStrategy = ot.MediumSafe(solver)
-algo = otexp.LineSampling(event, alpha, rootStrategy)
+algo = ot.LineSampling(event, alpha, rootStrategy)
 algo.setMaximumOuterSampling(1000)
 algo.setMaximumCoefficientOfVariation(5e-2)
 algo.run()
@@ -129,7 +128,7 @@ alpha_twoBranch = result.getStandardSpaceDesignPoint()
 ot.RandomGenerator.SetSeed(0)
 solver = ot.Brent(1e-3, 1e-3, 1e-3, 5)
 rootStrategy = ot.MediumSafe(solver)
-algo = otexp.LineSampling(event_twoBranch, alpha_twoBranch, rootStrategy)
+algo = ot.LineSampling(event_twoBranch, alpha_twoBranch, rootStrategy)
 algo.setSearchOppositeDirection(False)
 algo.setMaximumOuterSampling(1000)
 algo.setMaximumCoefficientOfVariation(5e-2)
@@ -148,7 +147,7 @@ print(result)
 ott.assert_almost_equal(result.getProbabilityEstimate(), 3e-4, 0.0, 1e-5)
 
 # LineSampling / system event / no adaptive alpha
-algo = otexp.LineSampling(unionEvent, alpha_twoBranch, rootStrategy)
+algo = ot.LineSampling(unionEvent, alpha_twoBranch, rootStrategy)
 ot.RandomGenerator.SetSeed(0)
 algo.setMaximumOuterSampling(1000)
 algo.setAdaptiveImportantDirection(False)
@@ -180,7 +179,7 @@ alpha = result.getStandardSpaceDesignPoint()
 # LineSampling
 rootStrategy = ot.SafeAndSlow(ot.Brent(1e-3, 1e-3, 1e-3, 5), 3, 0.1)
 ot.RandomGenerator.SetSeed(0)
-algo = otexp.LineSampling(event, alpha, rootStrategy)
+algo = ot.LineSampling(event, alpha, rootStrategy)
 algo.setMaximumOuterSampling(3000)
 algo.setMaximumCoefficientOfVariation(5e-2)
 algo.run()
@@ -191,7 +190,7 @@ ott.assert_almost_equal(result.getProbabilityEstimate(), 0.00110258, 0.0, 1e-4)
 # complementary
 event = ot.ThresholdEvent(Y, ot.Greater(), threshold)
 ot.RandomGenerator.SetSeed(0)
-algo = otexp.LineSampling(event, alpha, rootStrategy)
+algo = ot.LineSampling(event, alpha, rootStrategy)
[…]
