+++ python/doc/examples/data_analysis/sample_analysis/plot_quantile_confidence_exact_empirical.py
@@ -22,7 +22,6 @@ Compare the distribution of quantile estimators
 #
 # See  :ref:`quantile_confidence_estimation` and :ref:`quantile_asymptotic_confidence_estimation` to get theoretical details.
 import openturns as ot
-import openturns.experimental as otexp
 import openturns.viewer as otv
 
 
@@ -65,9 +64,9 @@ print("Empirical order statistics = ", k_emp)
 # %%
 # Now, we want to get the exact unilateral confidence interval that provides an upper bound of :math:`x_{\alpha}` with the confidence
 # :math:`\beta`, which means that :math:`\Prob{X_{(k_{up})} \geq x_{\alpha}} \geq \beta`.
-# This can be done using :class:`~openturns.experimental.QuantileConfidence`.
+# This can be done using :class:`~openturns.QuantileConfidence`.
 # We can get the order :math:`k_{up}` of the useful statistics. Care that the enumeration begins at 0.
-quantConf = otexp.QuantileConfidence(alpha, beta)
+quantConf = ot.QuantileConfidence(alpha, beta)
 k_up = quantConf.computeUnilateralRank(n)
 
 # %%
@@ -79,7 +78,7 @@ print("Upper bound order statistics = ", k_up)
 # %%
 # To get the exact unilateral confidence interval that provides a lower bound of :math:`x_{\alpha}` with the confidence
 # :math:`\beta`, which means that :math:`\Prob{X_{(k_{low})} \leq x_{\alpha}} \geq \beta`.
-# This can also be done using :class:`~openturns.experimental.QuantileConfidence`.
+# This can also be done using :class:`~openturns.QuantileConfidence`.
 # We can get the order :math:`k_{low}` of the useful statistics. Care: the enumeration begins at 0 here!
 k_low = quantConf.computeUnilateralRank(n, True)
 
