+++ python/doc/user_manual/data_analysis.rst
@@ -262,4 +262,4 @@ Quantile confidence
     :toctree: _generated/
     :template: class.rst_t
 
-    experimental.QuantileConfidence
+    QuantileConfidence
