+++ lib/src/Uncertainty/Algorithm/Simulation/CMakeLists.txt
@@ -21,7 +21,6 @@ ot_add_source_file (RandomDirection.cxx)
 ot_add_source_file (OrthogonalDirection.cxx)
 ot_add_source_file (PostAnalyticalControlledImportanceSampling.cxx)
 ot_add_source_file (PostAnalyticalImportanceSampling.cxx)
-ot_add_source_file (Wilks.cxx)
 ot_add_source_file (SubsetSamplingResult.cxx)
 ot_add_source_file (SubsetSampling.cxx)
 ot_add_source_file (QuadrantSampling.cxx)
@@ -41,7 +40,6 @@ ot_add_source_file (StandardSpaceCrossEntropyImportanceSampling.cxx)
 ot_add_source_file (PhysicalSpaceCrossEntropyImportanceSampling.cxx)
 
 ot_install_header_file (SimulationAlgorithm.hxx)
-ot_install_header_file (Wilks.hxx)
 ot_install_header_file (RootStrategyImplementation.hxx)
 ot_install_header_file (SamplingStrategy.hxx)
 ot_install_header_file (PostAnalyticalControlledImportanceSampling.hxx)
