### 2026-04-28 9d4e5af
LeastSquaresEquationsSolver: Moved out of experimental
