+++ python/doc/examples/calibration/mcmc_based_calibration/plot_imh_python_distribution.py
@@ -7,7 +7,7 @@ Sampling from an unscaled probability density
 # In this example, we show different ways to sample a distribution which PDF is known up to a normalization constant:
 #
 # - Case 1: computing the normalization factor and using a :class:`~openturns.PythonDistribution`,
-# - Case 2: using the Ratio of Uniforms algorithm with the class :class:`~openturns.experimental.RatioOfUniforms`
+# - Case 2: using the Ratio of Uniforms algorithm with the class :class:`~openturns.RatioOfUniforms`
 # - Case 3: using some Metropolis-Hastings algorithms with the class :class:`~openturns.IndependentMetropolisHastings`
 #   and :class:`~openturns.RandomWalkMetropolisHastings`.
 #
@@ -26,7 +26,6 @@ Sampling from an unscaled probability density
 # First, we draw the unscaled probability density function.
 
 import openturns as ot
-import openturns.experimental as otexp
 import openturns.viewer as otv
 from math import pi
 from time import monotonic
@@ -149,7 +148,7 @@ class NewDistribution_RoU(ot.PythonDistribution):
         )
         self.rangeLow_ = 0.0
         self.rangeUp_ = 2.0 * pi
-        self.sampler_ = otexp.RatioOfUniforms(self.logPDF_, self.getRange())
+        self.sampler_ = ot.RatioOfUniforms(self.logPDF_, self.getRange())
 
     def getRange(self):
         return ot.Interval(self.rangeLow_, self.rangeUp_)
@@ -203,12 +202,12 @@ print(f"Is Ratio of Uniforms faster ? {speedRoU > 10 * speedGeneric}")
 # %%
 # Case 2: Direct use the Ratio of Uniforms algorithm
 # --------------------------------------------------
-# In that case, we want to use the :class:`~openturns.experimental.RatioOfUniforms` algorithm to
+# In that case, we want to use the :class:`~openturns.RatioOfUniforms` algorithm to
 # sample :math:`p`. We need to compute the function :math:`\log f` and its range.
 #
-# We create the function :math:`\log f` and the :class:`~openturns.experimental.RatioOfUniforms`:
+# We create the function :math:`\log f` and the :class:`~openturns.RatioOfUniforms`:
 log_UnscaledPDF = ot.ComposedFunction(ot.SymbolicFunction("x", "log(x)"), f)
-ratio_algo = otexp.RatioOfUniforms(log_UnscaledPDF, range_PDF)
+ratio_algo = ot.RatioOfUniforms(log_UnscaledPDF, range_PDF)
 
 # %%
 # We can draw the exact PDF and the PDF estimated from a sample generated by the Ratio of Uniforms algorithm.
@@ -259,7 +258,7 @@ print("Acceptance ratio = ", ratio_algo.getAcceptanceRatio())
 
 # %%
 # We can estimate the normalization factor with the  Ratio of Uniforms algorithm (see
-# the documenttaion of :class:`~openturns.experimental.RatioOfUniforms`):
+# the documenttaion of :class:`~openturns.RatioOfUniforms`):
 print("Normalization factor estimate = ", ratio_algo.getC())
 
 # %%
