+++ python/test/t_RatioOfUniforms_std.py
@@ -1,10 +1,9 @@
 #! /usr/bin/env python
 
 import openturns as ot
-import openturns.experimental as otexp
 import openturns.testing as ott
 
-ratioOfU = otexp.RatioOfUniforms()
+ratioOfU = ot.RatioOfUniforms()
 print("ratioOfU is initialized?", ratioOfU.isInitialized())
 coll = ratioOfU.initialize()
 print("ratioOfU is initialized?", ratioOfU.isInitialized())
@@ -24,7 +23,7 @@ print("distribution = ", repr(distribution))
 
 
 # Now sample it using RoU
-ratioOfU = otexp.RatioOfUniforms(distribution)
+ratioOfU = ot.RatioOfUniforms(distribution)
 print("ratioOfU=", repr(ratioOfU))
 print("ratioOfU=", str(ratioOfU))
 print("ratioOfU is initialized?", ratioOfU.isInitialized())
@@ -44,12 +43,12 @@ print("One sample with try number", sample, "try number=", tryNumber)
 tri = ot.Triangular(0.0, 1.0, 1.0)
 ot.RandomGenerator.SetSeed(1234)
 logPDF = ot.SymbolicFunction("x", "log(x)")
-ratioOfU = otexp.RatioOfUniforms(logPDF, ot.Interval(0, 1))
+ratioOfU = ot.RatioOfUniforms(logPDF, ot.Interval(0, 1))
 print("Acceptance ratio=%.4g" % ratioOfU.getAcceptanceRatio())
 print("c=", ratioOfU.getC())
 sample1 = ratioOfU.getSample(10)
 ot.RandomGenerator.SetSeed(1234)
-ratioOfU = otexp.RatioOfUniforms(tri)
+ratioOfU = ot.RatioOfUniforms(tri)
 print("ratioOfU=", ratioOfU)
 sample2 = ratioOfU.getSample(10)
 ott.assert_almost_equal(sample1, sample2)
