### 2026-04-28 6424a6a
QuantileConfidence: Fix bugs, doc and add new methods.
Closes #3178
Closes #3182
Improve the performance of computeBilateralRank
Fix code for unilateral cases
Update documentation for quantile bounds description
Rename tail into lower_bounded to match doc
Add new coverage methods
