+++ python/test/t_QuantileConfidence_std.py
@@ -3,468 +3,738 @@
 import openturns as ot
 import openturns.testing as ott
 
-alpha = 0.05
-beta = 0.95
-algo = ot.QuantileConfidence(alpha, beta)
-print(algo)
-print(repr(algo))
-algo.setAlpha(alpha)
-algo.setBeta(beta)
-assert algo.getAlpha() == alpha
-assert algo.getBeta() == beta
 
-# lower rank
-print("lower rank ...")
-k_ref = {
-    59: 0,
-    93: 1,
-    124: 2,
-    153: 3,
-    181: 4,
-    208: 5,
-    311: 9,
-    410: 13,
-    506: 17,
-    717: 26,
-    809: 30,
-    900: 34,
-    1036: 40,
-}
-for i in range(991, 1013):
-    k_ref[i] = 38
-k_ref[1013] = 39
-for n in k_ref.keys():
-    ll = algo.computeUnilateralRank(n, True)
-    p = ot.Binomial(n, alpha).computeComplementaryCDF(ll)
-    print(f"n={n} l={ll} ref={k_ref[n]} p={p}")
-    assert ll == k_ref[n]
-    assert p >= beta
-
-# upper rank
-print("upper rank ...")
-algo.setAlpha(0.95)
-k_ref = {
-    59: 58,
-    93: 91,
-    124: 121,
-    153: 149,
-    181: 176,
-    208: 202,
-    311: 301,
-    410: 396,
-    506: 488,
-    601: 579,
-    717: 690,
-    809: 778,
-    900: 865,
-    1013: 973,
-    1036: 995,
-}
-for n in k_ref.keys():
-    u = algo.computeUnilateralRank(n)
-    p = ot.Binomial(n, alpha).computeCDF(u)
-    print(f"n={n} u={u} ref={k_ref[n]} p={p}")
-    assert u == k_ref[n]
-    assert p >= beta
-
-# bilateral ranks
-print("bilateral ranks ...")
-algo.setAlpha(0.05)
-algo = ot.QuantileConfidence(alpha, beta)
-k_ref = {
-    59: (0, 9),
-    60: (0, 8),
-    70: (0, 8),
-    80: (0, 8),
-    90: (0, 8),
-    100: (1, 10),
-    150: (1, 12),
-    200: (2, 15),
-    250: (3, 18),
-    300: (4, 21),
-    400: (11, 28),
-    500: (12, 33),
-    600: (19, 40),
-    700: (25, 50),
-    800: (22, 50),
-    900: (34, 69),
-    1000: (36, 63),
-    10000: (437, 536),
-    100000: (4879, 5160),
-}
-for n in k_ref.keys():
-    k1, k2 = algo.computeBilateralRank(n)
-    binomial = ot.Binomial(n, alpha)
-    p = binomial.computeProbability(ot.Interval(k1, k2))
-    print(f"n={n} k1={k1} k2={k2} p={p:.6f}")
-    assert (k1, k2) == k_ref[n]
-    assert p >= beta
-
-    # we can also make sure by exploring all combinations (costly)
-    if n > 100:
-        continue
-    bestP = 2
-    for i1 in range(n):
-        for i2 in range(i1, n):
-            # we compute P(k1<X<=k2)=P(k1+1<=X<=k2), CDF(k2)-CDF(k1) also works
-            p = binomial.computeProbability(ot.Interval(i1 + 1, i2))
-            if (p >= beta) and (p < bestP):
[…]
