+++ python/doc/developer_guide/release.rst
@@ -33,8 +33,15 @@ Release process
 
 12. Update `wheels <https://github.com/openturns/build-wheels>`_
 
-13. Set version X.Y+1dev, new changelog in master
+13. Update `doc <https://github.com/openturns/openturns.github.io>`_ latest link (copy folder X.Y into latest).
 
-14. Drop deprecated methods listed in TODO file
+14. Add discourse announcement topic.
 
-15. Bump soversion (lib/src/CMakeLists.txt)
+15. Update `website <https://github.com/openturns/www>`_, with latest doc link.
+
+16. Set version X.Y+1dev, new changelog in master
+
+17. Bump soversion (lib/src/CMakeLists.txt)
+
+18. Drop deprecated methods listed in TODO file
+
