### 2026-04-28 4610707
Analytical: Drop Starting point accessors
