### 2026-04-28 08a6df1
SWIG: Reorder modules
