+++ python/src/CMakeLists.txt
@@ -84,40 +84,6 @@ endmacro ()
 install (DIRECTORY openturns DESTINATION ${OPENTURNS_INCLUDE_PATH})
 
 
-ot_add_python_module(typ typ_module.i
-                      Collection.i Collection_doc.i
-                      PersistentCollection.i
-                      Description.i Description_doc.i
-                      PlatformInfo.i PlatformInfo_doc.i
-                      Point.i Point_doc.i
-                      PointWithDescription.i PointWithDescription_doc.i
-                      DomainImplementation.i DomainImplementation_doc.i
-                      Interval.i Interval_doc.i
-                      Indices.i Indices_doc.i
-                      IndicesCollection.i IndicesCollection_doc.i
-                      IndicesCollectionImplementation.i
-                      MatrixImplementation.i
-                      Matrix.i Matrix_doc.i
-                      SquareMatrix.i SquareMatrix_doc.i
-                      SymmetricMatrix.i SymmetricMatrix_doc.i
-                      Tensor.i Tensor_doc.i
-                      SymmetricTensor.i SymmetricTensor_doc.i
-                      ComplexTensor.i ComplexTensor_doc.i
-                      ComplexMatrix.i ComplexMatrix_doc.i
-                      HermitianMatrix.i HermitianMatrix_doc.i
-                      SquareComplexMatrix.i SquareComplexMatrix_doc.i
-                      TriangularMatrix.i TriangularMatrix_doc.i
-                      TriangularComplexMatrix.i TriangularComplexMatrix_doc.i
-                      CovarianceMatrix.i CovarianceMatrix_doc.i
-                      CorrelationMatrix.i CorrelationMatrix_doc.i
-                      IdentityMatrix.i IdentityMatrix_doc.i
-                      Sample.i Sample_doc.i
-                      HistoryStrategyImplementation.i HistoryStrategyImplementation_doc.i
-                      HistoryStrategy.i HistoryStrategy_doc.i
-                      BaseTypCollection.i
-                      BaseTypTemplateDefs.i
-                   )
-
 ot_add_python_module(func func_module.i
                       EnumerateFunction.i EnumerateFunction_doc.i
                       EnumerateFunctionImplementation.i EnumerateFunctionImplementation_doc.i
@@ -163,7 +129,7 @@ ot_add_python_module(func func_module.i
                       FieldToPointConnection.i FieldToPointConnection_doc.i
                       PointToFieldConnection.i PointToFieldConnection_doc.i
                       FieldToFieldConnection.i FieldToFieldConnection_doc.i
-		      VertexFieldToPointFunction.i VertexFieldToPointFunction_doc.i
+                      VertexFieldToPointFunction.i VertexFieldToPointFunction_doc.i
                       KarhunenLoeveProjection.i KarhunenLoeveProjection_doc.i
                       KarhunenLoeveLifting.i KarhunenLoeveLifting_doc.i
                       ValueFunction.i ValueFunction_doc.i
@@ -199,48 +165,6 @@ ot_add_python_module(func func_module.i
                       InverseTrendTransform.i InverseTrendTransform_doc.i
                       BaseFuncTemplateDefs.i
                    )
-ot_add_python_module(func2 func2_module.i
-                      ComposedHessian.i
-                      ConstantHessian.i ConstantHessian_doc.i
-                      DualLinearCombinationHessian.i DualLinearCombinationHessian_doc.i
-                      LinearCombinationHessian.i
-                      LinearGradient.i LinearGradient_doc.i
-                      MarginalHessian.i
-                      NoHessian.i
-                      NullHessian.i
-                      ParametricHessian.i
[…]
