+++ python/doc/user_manual/stochastic_process.rst
@@ -106,7 +106,7 @@ Gaussian process
     GaussianProcess
 
     :template: class.rst_t
-    experimental.ConditionedGaussianProcess
+    ConditionedGaussianProcess
 
 Functional basis process
 ========================
