### 2026-04-28 f066782
Linear/QuadraticLS: Remove deprecated method
