### 2026-04-29 #10484 @editorialbot
[REVIEW]: OpenTURNS for probabilistic modeling and uncertainty quantification
**Submitting author:** @mbaudin47 (<a href="http://orcid.org/0000-0002-0450-7947">Michael Baudin</a>)
**Repository:** https://github.com/openturns/openturns
**Branch with paper.md** (empty if default branch): josspaper
**Version:** v1.26
