### 2026-04-30 #10484 @tupui
My review was quite easy. I have known and used for my research OpenTURNS for what is now almost 10y (just realising how time is flying.) I did a few presentations myself on how to use OpenTURNS, had many discussions with the team. Wrote some tooling around it (e.g. Batman during my PhD which I made a JOSS paper on.)

I was never part of the core team or anything. As a maintainer of SciPy, I naturally am in collaboration with the broader Scientific Python Ecosystem. OpenTURNS is one of the good examples I happily cite. We have a good collaboration with the team on standards and they helped a lot for instance to bring Sobol' indices and QMC to SciPy (some personal contribution and Michael the author of the paper here was instrumental.) It's a very skilled team, kind, approachable and they have been maintaining the tools for a while building a real community around them. Kudos 👏

For the protocol, I did go through points individually now. I think the paper is also good. A bit on the longer side. But I personally would not bother asking to trim as I still believe papers here should be easy and if the authors felt like writing a bit more, then sure as long as the quality is there.

On the AI front, I can confirm the claims as I personally know the team and know their work ethic.

License: this is a LGPL and they have proper COPYING associated information. They do this very well here.

All in all, I really recommend for publication this paper. This would make for a great addition to the journal. I am glad we will now have a JOSS paper we will be able to use for citations.
