### 2026-04-30 #10484 @abhijit-c
OpenTURNS is an established, well-used, package in the UQ community. It is a very good fit for JOSS, and I do recommend accepting the submission. I've gone through each of the points carefully and believe that OpenTURNS has no major blockers from the software side.

I did manually verify that the software installs through pip and conda-forge. Additionally, I compiled it locally on two machines (osx-arm64 and linux-64) and verified its python and cpp test suite. Note, there is currently a failing test on osx-arm64 on the python side, but this is an issue with a downstream optional dependency (chaospy -> numpoly), and shouldn't be counted against OpenTURNS. I will file an issue with numpoly later; they're just missing osx-arm64 conda-forge builds, I can help with that. Otherwise, I see no issues with its claimed functionality and the test suite covers quite a bit.

The documentation is immense and comprehensive. It is well within the standard required of the journal.

I did have a couple of very minor notes on the paper itself that I believe need to be resolved

- Funding: I believe JOSS policy requires acknowledgement of funding sources for the project: https://joss.readthedocs.io/en/latest/submitting.html#policies . According to your about us page, OpenTURNS has had a few corporate sponsors, which would quality for this.
- AI Disclosure: The paper mentions generative AI use for editing purposes. I believe JOSS policy requires that the specific tool + version be named, according to: https://joss.readthedocs.io/en/latest/submitting.html#ai-usage-policy

In summary, definitely recommend for acceptance. Let me know when the paper is updated and I'll check off the last two boxes.

Reviewer AI Disclosure: I used Sonnet 4.6 to try and figure out whether the chaospy failing test was substantive, the results of which I verified myself.
