### 2026-05-02 2047871
Fix formatting in sensitivity_sobol_from_pce.rst
Fixes an error alread fixed in #2881, but introduced again in #2883.

The error is this: the input of the metamodel is X, not the standard input. The variance is not changed by an isoprobabilistic transform on the input. This is why the remaining equations, including the Sobol' indices are ok.
