### 2026-05-02 0ed9df1
paper/paper.md: Fixed AI disclosure
