+++ paper/paper.md
@@ -244,7 +244,7 @@ and knowledge exchange among researchers, engineers, and educators.
 # AI usage disclosure
 
 Generative AI was used to verify the grammar and spelling of this paper,
-check the choices of the words and check that the paper comply to the
+check the choices of the words and check that the paper comply with the
 requirements.
 No generative AI was used in the creation of the software, the technical 
 documentation, or the algorithm design.
