### 2026-05-02 a57db32
paper/paper.md: Take review into account
- Define the name and version of the generative AI tool
- Created a new Acknowledgements section to explain funding
- Fixed SALOME case.
- Improved the "State of the field" section.
- paper/paper.bib: Fixed formatting
