+++ paper/paper.bib
@@ -1,148 +1,316 @@
-@Inbook{baudin2016,
-  author={Baudin, Micha{\"e}l and Dutfoy, Anne and Iooss, Bertrand and Popelin, Anne-Laure},
-  title={{OpenTURNS: An Industrial Software for Uncertainty Quantification in Simulation. In: Ghanem R., Higdon D., Owhadi H. (eds) Handbook of Uncertainty Quantification.}},
-  year={2016},
-  publisher={Springer International Publishing},
-  address={Cham},
-  pages={1--38},
-  isbn={978-3-319-11259-6},
-  doi={10.1007/978-3-319-11259-6_64-1},
-  url={https://doi.org/10.1007/978-3-319-11259-6_64-1}
+@inbook{baudin2016,
+	title        = {{OpenTURNS: An Industrial Software for Uncertainty Quantification in Simulation}},
+	author       = {Baudin, Micha{\"e}l and Dutfoy, Anne and Iooss, Bertrand and Popelin, Anne-Laure},
+	year         = 2016,
+	booktitle    = {Handbook of Uncertainty Quantification},
+	publisher    = {Springer International Publishing},
+	address      = {Cham},
+	pages        = {1--38},
+	doi          = {10.1007/978-3-319-11259-6_64-1},
+	isbn         = {978-3-319-11259-6},
+	url          = {https://doi.org/10.1007/978-3-319-11259-6_64-1},
+	editor       = {Ghanem, Roger and Higdon, David and Owhadi, Houman}
 }
-
 @article{dutfoy2023,
-  author = {A. Dutfoy},
-  year = {2023},
-  title = {Uncertainty on estimated magnitudes: a new approach based on a {Poisson} point process of dimension 2},
-  journal = {Pure Appl. Geophys.},
-  volume = {180},
-  number = {10},
-  pages = {919-933},
-  url = {https://doi.org/10.1007/s00024-022-03222-6},
-  doi = {10.1007/s00024-022-03222-6}
+	title        = {Uncertainty on estimated magnitudes: a new approach based on a {Poisson} point process of dimension 2},
+	author       = {Dutfoy, Anne},
+	year         = 2023,
+	journal      = {Pure and Applied Geophysics},
+	volume       = 180,
+	number       = 10,
+	pages        = {919--933},
+	doi          = {10.1007/s00024-022-03222-6},
+	url          = {https://doi.org/10.1007/s00024-022-03222-6}
 }
-
 @misc{openturns_forum,
-  title        = {{OpenTURNS}'s forum},
-  howpublished = {\url{https://openturns.discourse.group/}},
-  year = {2020}
+	title        = {{OpenTURNS}'s forum},
+	year         = 2020,
+	howpublished = {\url{https://openturns.discourse.group/}}
 }
-
 @misc{persalys,
-  title = {{Persalys: Graphical User Interface for Uncertainty Quantification}},
-  author = {{EDF} and {Phimeca Engineering}},
-  year = {2016},
-  howpublished = {\url{https://www.persalys.fr/}}
+	title        = {{Persalys: Graphical User Interface for Uncertainty Quantification}},
+	author       = {{EDF} and {Phimeca Engineering}},
+	year         = 2016,
+	howpublished = {\url{https://www.persalys.fr/}}
 }
-
 @article{pedregosa2011scikit,
-  title={{Scikit-learn: Machine learning in Python}},
-  author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
-  journal={The Journal of Machine Learning Research},
-  volume={12},
-  pages={2825--2830},
[…]
