### 2026-05-02 2396ff0
paper/paper.md: Define the name of the generative AI tool
