### 2026-05-02 719880e
paper/paper.md: Created a new Acknowledgements section to explain funding
