### 2026-05-02 #10484 @mbaudin47
@tupui : Thank you for your thorough review.

@abhijit-c For the AI disclosure topic, [2396ff0d](https://github.com/openturns/openturns/pull/3181/changes/2396ff0dd0a0a4c439b86fd766e1515e64024f8b) fixes that issue. I also fixed a typo mentioned at the pre-review stage by @Nikoleta-v3 in [0ed9df12](https://github.com/openturns/openturns/pull/3181/changes/0ed9df12d1ad4b89bc7ac76897fc1b93baf53a64).

Regarding the funding, you are absolute right on this point: the paper is too weak on this topic. Thank you very much for pointing this. The development is mainly done directly by the members of the OpenTURNS consortium, with a few contributions from other developers. From time to time, there are contracts where members of the consortium (mainly EDF and Airbus) fund developments by external companies. Commit [719880ece0](https://github.com/openturns/openturns/pull/3181/changes/719880ece0b75a2c70338596a8f47f3c930b37e9) updates the manuscript to reflect this. To clarify the process, let me give two examples: the development in PR [2422](https://github.com/openturns/openturns/pull/2422) was specified by EDF and performed by @ariste-sarl, and PR [1574](https://github.com/openturns/openturns/pull/1574) was specified by EDF and performed by @KieranDelamotte. I hope that the development process is clear enough and that the paper update is ok from your point of view.

When the review process is more advanced, I will move the PR [3181](https://github.com/openturns/openturns/pull/3181) from draft to ready and make it be merged.
