### 2026-05-06 16461ad
Improved LevelSetMesher
Now one can mesh a LevelSet starting from a field. It allows to
start from an existing mesh more suited to describe the level set than its bounding box (in particular in high dimension) and
it also allows to reuse already computed
values if one want to mesh level sets associated to a given
function but for different levels.
