### 2026-05-06 6e22849
Dlib: Workaround gcc 16 warning
