### 2026-05-06 3ee232f
BlockIndependent: Unfold lists
