### 2026-05-06 f0dad0f
BlockIndependent: Unfold lists
