### 2026-05-15 openturns/openturns#3196 @dannys4
Technically not-independent Gaussian mixture sampling
### Describe the issue

The documentation nominally says that taking two sets of samples of prescribed sizes and mixing them constitutes independent sampling from a Gaussian mixture---this is _technically_ incorrect in the finite sample case.

### Where can the issue be found?

https://openturns.github.io/openturns/master/auto_data_analysis/distribution_fitting/plot_smoothing_mixture.html

### Additional Context

I acknowledge that this is _extremely_ pedantic---I think even a stray remark to the effect of "this sampling procedure is biased for small $n$" would fix the problem, and there's no need to fully re-engineer the example (I did check that a mixture model in `openturns` C++ does indeed sample correctly).

I am largely filing this w.r.t. [JOSS Review](https://github.com/openjournals/joss-reviews/issues/10484) to show that I'm looking relatively closely at the documentation and have, so far, only found a single minor pedagogical mistake.
