### 2026-05-16 openturns/openturns#3196 @mbaudin47
Do you refer to this particular sentence?

> In this section, we show that a mixture of two Normal distributions is nothing more than the merged sample of two independent Normal distributions.
