### 2026-05-16 openturns/openturns#3196 @dannys4
yeah, for user-specified $n$, just concatenating a proportionate number of samples is _technically_ not independently sampling from the distribution (e.g., consider the case where $w_1=0.01$ and $w_2=0.99$ and $n=99$---what's in the documentation would suggest we never sample from the first mode). For $n\geq 1000$ in this case, however, there's really not a meaningful difference. It's just an incredibly pedantic nitpick that I think is often glossed over in material intended for learning about UQ and Monte Carlo.

For pedagogical reasons, I should also mention the case of $w_1=w_2=0.5$ and $n=2$---obviously, after sampling the first point, your second point is clearly not independent from the first since the second point _must_ be sampled from the second mixture mode.
