### 2026-05-16 #10484 @dannys4
Yeah, I don't envy anyone who has to list every relevant software for an entire field over 20 years. This is a valuable comment in- and of- itself. Thanks!
