### 2026-05-17 openturns/openturns#3196 @mbaudin47
You are absolutely right. Thank you for pointing this error. This is not pedantic: the content is just wrong. I wrote the main part of the example. But another developer was in charge of introducing into the doc and introduced the faulty section. Then the important difference between the infinite and the finite sample cases was not noticed. I just took the (Wand & Jones, 1994) book from my shelf and read the paragraph. Notice that the expression from the original book is as confusing as our own help page and does not state the important fact that this property is true only for the population, not in the finite sample case.

## Generate an observation from a mixture of two distribution

Before analyzing the help page, the maths must be clarified. Here is a method to sample from the mixture distribution $w_1 f_1 + w_2 f_2$ where $w_1, w_2 \geq 0$ and $w_1 + w_2 = 1$.

- Generate $u$ from the distribution $\mathcal{U}(0,1)$.
- If $u \leq w_1$, then generate $x$ from $f_1$.
- Otherwise, generate $x$ from $f_2$.

On output, the real number $x$ is an observation from the mixture.

Hence, when the sample size $n \to \infty$, there is a fraction $w_1$ of the population which has the distribution $f_1$ and the remaining fraction $w_2$ from $f_2$.

## A review of the errors in the example

* **Mathematical error.** The statement that exactly 1/4 of the observations come from the first distribution and 3/4 from the second is mathematically incorrect. This holds true only as $n$ tends to infinity. Consequently, the section "[generate-the-mixture-by-merging-two-samples](https://openturns.github.io/openturns/master/auto_data_analysis/distribution_fitting/plot_smoothing_mixture.html#generate-the-mixture-by-merging-two-samples)" must be completely removed. First, it is outside the scope of the example: the goal of the example is to show the sensitivity of kernel smoothing to the bandwidth, not the properties of a mixture distribution. Second, it is mathematically incorrect and pedagogically confusing.

* **Inconsistent notation.** There is an inconsistency in the definition of the mixture distribution. It is written as $w_1 f_A(x) + w_2 f_B(x)$, whereas a more consistent form would be $w_1 f_1(x) + w_2 f_2(x)$. In other words, the weight $w_1$ should correspond to the distribution $f_1$ and the weight $w_2$ to the distribution $f_2$. This notation was chosen because (Wand & Jones, 1994) chose the letter $f_1$ to introduce the mixture, but with different notations for the individual distributions.

* **Inversion of weights and text.** There is a contradiction between the formula and the description. The formula states $w_1 = \frac{3}{4}$ for $f_A$ and $w_2 = \frac{1}{4}$ for $f_B$. However, the text states that "1/4th of the observations have the $\mathcal{N}(0,1)$ distribution $f_A$ and 3/4th of the observations have the $\mathcal{N}(3/2,(1/3)^2)$ distribution $f_B$." The weights are inverted in the prose.
