### 2026-05-17 bd41f91
Clarify distribution definitions and sampling comments
Closes #3196

Added comments to clarify sample size and distribution definitions.
Refactor variable names for clarity and consistency
