### 2026-05-17 #10484 @mbaudin47
@dannys4 : In [fd1b5b39876c](https://github.com/openturns/openturns/pull/3181/changes/fd1b5b39876ce0b90b60d5bf6a2ba519d2023cb2), I reworked the "State of field" section. Can you please review this update and see if the updated version matches your requirements?
