### 2026-05-17 #10484 @dannys4
yeah that looks like well more than enough. Again, sorry to make you jump through hoops for the paper---I did mean to just contextualize the problems, and I'll definitely be more clear (and probably less nit-picky) in the future. Thanks @tupui for a review review :)
