### 2026-08-18 #10484 @crvernon
@mbaudin47 - Please address the following before I accept for publication:

In your archive:

- [ ] The author order and included authors in your Zenodo archive must math those listed in your paper exactly.  Please edit the metadata of your Zenodo archive to reflect this.  No new release required.

In your paper:

- [ ] LINE 32: “Uncertainty Quantification” should be lowercase here: “uncertainty quantification,” since it is not a proper noun
- [ ] LINE 47: “Gaussian Processes” should be written as “Gaussian processes”
- [ ] LINES 68-70: Avoid nested citation parentheses. For example, change “SALOME (see (SALOME Consortium, 2026))” to “SALOME (see SALOME Consortium, 2026)” or simply “SALOME (SALOME Consortium, 2026).” Make the same change for pSeven.
- [ ] LINES 142-143: The sentence is missing the names of the cited tools: “including (IRT Saint Exupéry, 2026) and (P. T. Roy et al., 2018).” Identify them directly, e.g., “including GEMSEO (IRT Saint Exupéry, 2026) and BATMAN (P. T. Roy et al., 2018)”
- [ ] LINES 251-253: The bibliography should not display the literal author text “& others.” Replace this with the correctly encoded author list in your bib file.

Let me know when these changes have been addressed.
