### 2026-08-31 #10484 @mbaudin47
> [@mbaudin47](https://github.com/mbaudin47) just checking in to see if you have had time to address the above?

- The paper was updated and the PR was merged a few hours ago (see https://github.com/openturns/openturns/pull/3283).
- The https://zenodo.org/records/21060713 archive has been updated

Thanks to @jschueller :)

Is this ok from you point of view?
