### 2026-09-02 #10484 @crvernon
🥳 Congratulations on your new publication @mbaudin47! Many thanks to @Nikoleta-v3 for editing and @abhijit-c, @dannys4, and @tupui for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
