### 2026-09-02 #10484 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
