### 2026-09-03 #10484 @mbaudin47
Thanks to you @crvernon and to other reviewers as well: the OpenTURNS developers really appreciate your work and the journal!
